% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_deterministics.R
\name{add_deterministics}
\alias{add_deterministics}
\title{Add Deterministic Terms to Country Models}
\usage{
add_deterministics(object, const = NULL, trend = NULL, seasonal = NULL)
}
\arguments{
\item{object}{a named list, usually, the output of a call to \code{\link{create_models}}.}

\item{const}{either logical for VAR models or a character specifying whether a constant term
enters the error correction term (\code{"restricted"}) or the non-cointegration term (\code{"unrestricted"})
of a VEC model. If `NULL` (default) no constant term will be added.}

\item{trend}{either logical for VAR models or a character specifying whether a trend term
enters the error correction term (\code{"restricted"}) or the non-cointegration term (\code{"unrestricted"})
of a VEC model. If `NULL` (default) no trend term will be added.}

\item{seasonal}{either logical for VAR models or a character specifying whether seasonal dummies
enter the error correction term (\code{"restricted"}) or the non-cointegreation term (\code{"unrestricted"}).
If `NULL` (default) no seasonal terms will be added. The amount of dummy variables depends
on the frequency of the time series data in each list element.}
}
\value{
A list of country models.
}
\description{
Adds deterministic terms to a list of country models, which was produced by
the function \code{\link{create_models}}.
}
\examples{
data("gvar2016")

country_data <- gvar2016$country_data
global_data <- gvar2016$global_data
region_weights <- gvar2016$region_weights
weight_data <- gvar2016$weight_data

# Take first difference of all variables y and Dp
country_data <- diff_variables(country_data, variables = c("y", "Dp", "r"))

# Generate EA area region with 2 year rolling window weights
ea <- c("AT", "BE", "DE", "ES", "FI", "FR", "IT", "NL")
temp <- create_regions(country_data = country_data,
                       regions = list("EA" = ea),
                       period = 2,
                       region_weights = region_weights,
                       weight_data = weight_data)
country_data <- temp$country_data
weight_data <- temp$weight_data

# Generate weight matrices as 2 year rolling window averages
gvar_weights <- create_weights(weight_data = weight_data, period = 2,
                               country_data = country_data)


# Create an object with country model specifications
model_specs <- create_specifications(country_data = country_data,
                                     global_data = global_data,
                                     variables = c("y", "Dp", "r"),
                                     countries = c("EA", "US", "JP", "CA", "MX", "GB"),
                                     p_domestic = 1,
                                     p_foreign = 1,
                                     type = "VAR")

# Create estimable objects
object <- create_models(country_data = country_data,
                        gvar_weights = gvar_weights,
                        model_specs = model_specs)

# Add deterministics
object <- add_deterministics(object, const = TRUE)

# Add priors
object <- add_priors(object)


}
