% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_posterior_coefficients.gvarmodel.R
\name{add_posterior_coefficients.gvarmodel}
\alias{add_posterior_coefficients.gvarmodel}
\title{Posterior Simulation of Model Coefficients}
\usage{
\method{add_posterior_coefficients}{gvarmodel}(object, ...)
}
\arguments{
\item{object}{an object of class 'gvarmodel', usually, a result of a
call to \code{\link{create_gvarmodel}} in combination with
\code{\link[bvartools]{add_priors}} and \code{\link[bvartools]{add_initial_values}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list of class 'gvarmodel'.
}
\description{
Forwards model input to posterior simulation functions for vector autoregressive models.
}
