% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_posterior_coefficients.gvecmodel.R
\name{add_posterior_coefficients.gvecmodel}
\alias{add_posterior_coefficients.gvecmodel}
\title{Posterior Simulation of Model Coefficients}
\usage{
\method{add_posterior_coefficients}{gvecmodel}(object, ...)
}
\arguments{
\item{object}{an object of class 'gvecmodel', usually, a result of a
call to \code{\link{create_gvecmodel}} in combination with
\code{\link[bvartools]{add_priors}} and \code{\link[bvartools]{add_initial_values}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list of class 'gvecmodel'.
}
\description{
Forwards model input to posterior simulation functions for vector error
correction models.
}
