% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_posterior_loglik.gvarmodel.R
\name{add_posterior_loglik.gvarmodel}
\alias{add_posterior_loglik.gvarmodel}
\title{Add Log-Likelihood}
\usage{
\method{add_posterior_loglik}{gvarmodel}(object, ...)
}
\arguments{
\item{object}{an object of class 'gvarmodel'.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list of class 'gvarmodel'.
}
\description{
Calculates and adds posterior log-likelihoods to the sub-models of an object
of class 'gvarmodel'.
}
