% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_posterior_loglik.gvecmodel.R
\name{add_posterior_loglik.gvecmodel}
\alias{add_posterior_loglik.gvecmodel}
\title{Add Log-Likelihood}
\usage{
\method{add_posterior_loglik}{gvecmodel}(object, ...)
}
\arguments{
\item{object}{an object of class 'gvecmodel'.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list of class 'gvecmodel'.
}
\description{
Calculates and adds posterior log-likelihoods to the sub-models of an object
of class 'gvecmodel'.
}
