% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_priors.gvarmodel.R
\name{add_priors.gvarmodel}
\alias{add_priors.gvarmodel}
\title{Add Priors to Bayesian Models}
\usage{
\method{add_priors}{gvarmodel}(object, ...)
}
\arguments{
\item{object}{a list of models, usually, the output of a call to
\code{\link{create_gvarmodel}} in combination with \code{\link{add_submodels}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list of class 'gvarmodel'.
}
\description{
Adds prior specifications to a list of models by passing each element to
the respective method.
}
