% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_priors.gvecmodel.R
\name{add_priors.gvecmodel}
\alias{add_priors.gvecmodel}
\title{Add Priors to Bayesian Models}
\usage{
\method{add_priors}{gvecmodel}(object, ...)
}
\arguments{
\item{object}{a list of models, usually, the output of a call to
\code{\link{create_gvecmodel}} in combination with \code{\link{add_submodels}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list of class 'gvecmodel'.
}
\description{
Adds prior specifications to a list of models by passing each element to
the respective method.
}
