% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_priors.R
\name{add_priors}
\alias{add_priors}
\title{Add Priors to Country Models}
\usage{
add_priors(
  object,
  coef = list(v_i = 1, v_i_det = 0.1),
  coint = list(v_i = 0, p_tau_i = 1),
  sigma = list(df = 3, scale = 1),
  ssvs = list(inprior = 0.5, tau = c(0.05, 10)),
  bvs = list(inprior = 0.5)
)
}
\arguments{
\item{object}{a named list, usually, the output of a call to \code{\link{create_models}}.}

\item{coef}{a named list of prior specifications for the coefficients of the country-specific
models. For the default specification all prior means are set to zero and the diagonal elements of
the prior variance-covariance matrix are set to 1 for coefficients corresponding to non-deterministic
terms. For deterministic coefficients the prior variances are set to 10. The variances need to be
specified as precisions, i.e. as inverses of the variances. For further specifications see 'Details'.}

\item{coint}{a named list of prior specifications for cointegration coefficients of the
country-specific models. See 'Details'.}

\item{sigma}{a named list of prior specifications for the error variance-covariance matrix
of the country-specific models. For the default specification of an inverted Wishart distribution
the prior degrees of freedom are set to 3 and the prior variances to 1. See 'Details'.}

\item{ssvs}{a named list of prior specifications for the SSVS algorithm. See 'Details'.}

\item{bvs}{a named list of prior specifications for the BVS algorithm. See 'Details'.}
}
\value{
A list of country models.
}
\description{
Adds prior specifications to a list of country models, which was produced by
the function \code{\link{create_models}}.
}
\details{
The argument \code{coef} can contain the following elements
\describe{
  \item{\code{v_i}}{a numeric specifying the prior precision of the coefficients. Default is 1.}
  \item{\code{v_i_det}}{a numeric specifying the prior precision of coefficients corresponding to deterministic terms. Default is 0.1.}
  \item{\code{coint_var}}{a logical specifying whether the prior mean of the first own lag of an
  endogenous variable should be set to 1. Default is \code{FALSE}.}
  \item{\code{kappa}}{a numeric vector of length 4 containing parameters for the calculation of
  the Minnesota prior. For the endogenous variable \eqn{i} the prior variance of the \eqn{l}th
  lag of regressor \eqn{j} is obtained as
  \deqn{ \frac{\kappa_{1}}{l^2} \textrm{ for own lags of domestic endogenous variables,}}
  \deqn{ \frac{\kappa_{2}}{l^2} \frac{\sigma^2_{i}}{\sigma^2_{j}} \textrm{ for domestic endogenous variables other than own lags,}}
  \deqn{ \frac{\kappa_{3}}{(l + 1)^2} \frac{\sigma^2_{i}}{\sigma^2_{j}} \textrm{ for foreign and global exogenous variables,}}
  \deqn{ \kappa_{4} \sigma^2_{i}  \textrm{ for deterministic terms.}}
  The indices of \eqn{\kappa} correspond to the positions of the elements in the argument \code{kappa}.
  \eqn{\sigma_{i}} is the residual standard deviation of variable \eqn{i} of an unrestricted
  OLS estimate of the model. For exogenous variables \eqn{\sigma_{i}} corresponds to the standard
  deviation of the original series.
  
  For VEC models the function only provides priors for the non-cointegration part of the model. The
  residual standard errors \eqn{\sigma_i} are based on an unrestricted OLS regression of the
  endogenous variables on the error correction term and the non-cointegration regressors.}
  \item{\code{max_var}}{numeric specifying the maximum prior variance that is allowed for non-deterministic coefficients.}
}
If \code{kappa} is specified, \code{v_i} and \code{v_i_det} will be ignored.

The argument \code{coint} can contain the following elements
\describe{
  \item{\code{coint_v_i}}{numeric between 0 and 1 specifying the shrinkage of the cointegration space prior. Default is 0.}
  \item{\code{coint_p_tau_i}}{numeric of the diagonal elements of the inverted prior matrix of
the central location of the cointegration space \eqn{sp(\beta)}. Default is 1.}
}

The argument \code{sigma} can contain the following elements
\describe{
  \item{\code{df}}{an integer specifying the prior degrees of freedom of the error term.
  Default is 3. If a VEC model is estimated, the rannk \eqn{r} of the cointegration matrix
  is automatically added to \code{df}.}
  \item{\code{scale}}{a numeric specifying the prior error variance of the endogenous variables.
  Default is 1.}
}

The argument \code{ssvs} can contain the following elements
\describe{
  \item{\code{inprior}}{a numeric between 0 and 1 specifying the prior probability
  of a variable to be included in the model. Default is 0.5.}
  \item{\code{tau}}{a numeric vector of two elements containing the prior standard errors
  of restricted variables (\eqn{\tau_0}) as its first element and unrestricted variables (\eqn{\tau_1})
  as its second. Default is \code{c(0.05, 10)}.}
  \item{\code{semiautomatic}}{an numeric vector of two elements containing the
  factors by which the standard errors associated with an unconstrained least squares
  estimate of the country VARX model are multiplied to obtain the prior standard errors
  of restricted (\eqn{\tau_0}) and unrestricted (\eqn{\tau_1}) variables. This is the
  semiautomatic approach described in George et al. (2008).}
}
Either \code{tau} or \code{semiautomatic} must be specified.

The argument \code{bvs} can contain the following elements
\describe{
  \item{\code{inprior}}{a numeric between 0 and 1 specifying the prior probability
  of a variable to be included in the model. Default is 0.5.}
  \item{\code{kappa}}{a numeric vector of length 4 containing parameters for the calculation of
  the Minnesota-like inclusion priors, where
  \deqn{ \kappa_{1} \textrm{ for own lags of domestic endogenous variables,}}
  \deqn{ \frac{\kappa_{2}}{l} \textrm{ for domestic endogenous variables other than own lags,}}
  \deqn{ \frac{\kappa_{3}}{l + 1} \textrm{ for foreign and global exogenous variables,}}
  \deqn{ \kappa_{4} \textrm{ for deterministic terms.}}
  The indices of \eqn{\kappa} correspond to the positions of the elements in the argument \code{kappa}.}
}
}
\examples{
data("gvar2016")

country_data <- gvar2016$country_data
global_data <- gvar2016$global_data
region_weights <- gvar2016$region_weights
weight_data <- gvar2016$weight_data

# Take first difference of all variables y and Dp
country_data <- diff_variables(country_data, variables = c("y", "Dp", "r"))

# Generate EA area region with 2 year rolling window weights
ea <- c("AT", "BE", "DE", "ES", "FI", "FR", "IT", "NL")
temp <- create_regions(country_data = country_data,
                       regions = list("EA" = ea),
                       period = 2,
                       region_weights = region_weights,
                       weight_data = weight_data)
country_data <- temp$country_data
weight_data <- temp$weight_data

# Generate weight matrices as 2 year rolling window averages
gvar_weights <- create_weights(weight_data = weight_data, period = 2,
                               country_data = country_data)

# Create an object with country model specifications
model_specs <- create_specifications(country_data = country_data,
                                     global_data = global_data,
                                     variables = c("y", "Dp", "r"),
                                     countries = c("EA", "US", "JP", "CA", "MX", "GB"),
                                     p_domestic = 1,
                                     p_foreign = 1,
                                     type = "VAR")

# Create estimable objects
object <- create_models(country_data = country_data,
                        gvar_weights = gvar_weights,
                        model_specs = model_specs)

# Add priors
object <- add_priors(object)

}
\references{
Chan, J., Koop, G., Poirier, D. J., & Tobias J. L. (2019). \emph{Bayesian econometric methods}
(2nd ed.). Cambridge: Cambridge University Press.

George, E. I., Sun, D., & Ni, S. (2008). Bayesian stochastic search for VAR model
restrictions. \emph{Journal of Econometrics, 142}(1), 553--580.
\url{https://doi.org/10.1016/j.jeconom.2007.08.017}

Korobilis, D. (2013). VAR forecasting using Bayesian variable selection.
\emph{Journal of Applied Econometrics, 28}(2), 204--230.
}
