% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_submodels.R
\name{add_submodels}
\alias{add_submodels}
\title{Add Sub-Models to Global Model
 
Generic function used to generate sub-models and add them to a model object.}
\usage{
add_submodels(object, ...)
}
\arguments{
\item{object}{an object of a class, for which a method should be called.}

\item{...}{arguments passed forward to method.}
}
\description{
Add Sub-Models to Global Model
 
Generic function used to generate sub-models and add them to a model object.
}
