% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_weight_matrices.gvarmodel.R
\name{add_weight_matrices.gvarmodel}
\alias{add_weight_matrices.gvarmodel}
\title{Add Weight Matrices}
\usage{
\method{add_weight_matrices}{gvarmodel}(object, submodel_data, period)
}
\arguments{
\item{object}{an object of class 'gvarmodel'. Usually, the output of a call
to \code{\link{create_gvarmodel}}.}

\item{submodel_data}{a list of class 'submodeldata' consisting of two
time-series objects, \code{endogen} and \code{weights}, which contain
sub-model-specific observations of endogenous variables and weights, respectively.}

\item{period}{either an integer for time varying weights or a numeric
vector specifying the periods in \code{weights} that
should be used to calculate constant weights. See 'Details'.}
}
\value{
An object of class 'gvarmodel'.
}
\description{
Adds weight matrices for each sub-model in an object of class 'gvarmodel'.
}
\details{
The function creates of sub-model-specific weight matrices.
If a numeric vector is provided as \code{period}, the function calculates
weights based on the sums over the specified periods. If an integer
is proved, the weights are constructed from rolling sums over the last
\code{period} periods. If a sub-model series begins earlier than its
corresponding weight series, the sums over the first \code{period}
observations of a sub-model's weight data are used until the periods match.
}
\examples{
# Load data
data("gvar2023")
submodel_data <- gvar2023[["submodel_data"]] # Country series including weights
global_data <- gvar2023[["global_data"]] # Global commodities data

# Create 'gvarmodel' object
object <- create_gvarmodel(submodel_data = submodel_data,
                           global_data = global_data)

# Generate weight matrices as 3 year, rolling window averages
object <- add_weight_matrices(object = object,
                              submodel_data = submodel_data,
                              period = 3)


}
