% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_model_obs.gvarmodel.R
\name{align_model_obs.gvarmodel}
\alias{align_model_obs.gvarmodel}
\title{Align Observations across Models}
\usage{
\method{align_model_obs}{gvarmodel}(object, ...)
}
\arguments{
\item{object}{an object of class 'gvarmodel'.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object of class 'gvarmodel'.
}
\description{
Restricts each sub-model in an object of class 'gvarmodel' to the set of
observations common to all sub-models, ensuring that comparisons are computed
on the same underlying sample.
}
