% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_model_obs.gvecmodel.R
\name{align_model_obs.gvecmodel}
\alias{align_model_obs.gvecmodel}
\title{Align Observations across Models}
\usage{
\method{align_model_obs}{gvecmodel}(object, ...)
}
\arguments{
\item{object}{an object of class 'gvecmodel'.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object of class 'gvecmodel'.
}
\description{
Restricts each sub-model in an object of class 'gvecmodel' to the set of
observations common to all sub-models, ensuring that comparisons are computed
on the same underlying sample.
}
