% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgvars.R
\docType{package}
\name{bgvars}
\alias{bgvars}
\title{bgvars: Bayesian Inference for Global Vector Autoregressive (GVAR) and Global Vector Error Correction (GVEC) Models}
\description{
R/C++ implementation of Bayesian global vector autoregressive and global vector error correction models.
}
\author{
Franz X. Mohr
}
