% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvec_to_bvar.R
\name{bvec_to_bvar.vecxsubmodelest}
\alias{bvec_to_bvar.vecxsubmodelest}
\title{Transform a VEC Model to a VAR in Levels}
\usage{
bvec_to_bvar.vecxsubmodelest(object)
}
\arguments{
\item{object}{an object of class 'vecxsubmodelest'.}
}
\value{
An object of class 'varxsubmodelest'.
}
\description{
An object of class 'vecxsubmodelest' is transformed to a VAR in levels.
}
\references{
Lütkepohl, H. (2006). \emph{New introduction to multiple time series analysis} (2nd ed.). Berlin: Springer.
}
