% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_best_model.submodelselcritlist.R
\name{choose_best_model.submodelselcritlist}
\alias{choose_best_model.submodelselcritlist}
\title{Choose Best Model}
\usage{
\method{choose_best_model}{submodelselcritlist}(object, criterion = "BIC", ...)
}
\arguments{
\item{object}{object of class 'submodelselcritlist', usually, a result of a call
to \code{\link[bvartools]{selection_criteria}}.}

\item{criterion}{the selection criterion that should be plotted. Available choices
are \code{"LL"}, \code{"AIC"}, \code{"BIC"} (default), \code{"HQ"}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A vector of integers containing the positions of the best models
in the list provided in argument \code{object}.
}
\description{
Chooses the best model according the selection criteria in an object of class
'submodelselcritlist'.
}
\details{
If argument \code{criterion} is \code{"LL"}, the model with the maximum value is chosen,
otherwise, the model with the minimum value.
}
