% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_submodels.R
\name{choose_submodels}
\alias{choose_submodels}
\title{Susbmodel Selection}
\usage{
choose_submodels(object, ic = "BIC", select = "order")
}
\arguments{
\item{object}{a list containing the results of the country estimates, usually, the
result of a call to \code{\link{draw_posterior}}.}

\item{ic}{a character specifying the information criterion used for model selection.
Available options are "AIC", "BIC" (default) and "HQ".}

\item{select}{a character specifying how the best country model is selected.
If \code{"order"} (default), the country model with the overall minimum value of the
specified information criterion per country is selected. If \code{"rank"}, the
function selects the model, after which the selected information criterion increases
for the first time.}
}
\description{
Selects the best country-specific submodel of a GVAR model.
}
