% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contemp_coeffs.R
\name{contemp_coeffs}
\alias{contemp_coeffs}
\title{Contemporaneous Coefficients}
\usage{
contemp_coeffs(
  object,
  type = "mean",
  tt = NULL,
  all = FALSE,
  variables = NULL,
  ic = "BIC",
  select = "order"
)
}
\arguments{
\item{object}{a list containing the results of the country estimates, usually, the
result of a call to \code{\link{estimate_gvar}}.}

\item{type}{a character specifying the type of statistic that should be calculated.
Possible option are \code{"mean"} (default), \code{"median"} and \code{"sd"}.}

\item{tt}{a numeric specifying the time index of the coefficients. Only applicable if the country models were
estimated with time-varying parameters or stochastic volatility.
If \code{NULL} the latest or only available period will be used.}

\item{all}{a logical. If \code{TRUE}, the contemporaneous effects of all estimated models will be printed. Otherwise, only
those models will be considered, which enter the global model.}

\item{variables}{a character vector specifying the variables, for which contemporaneous coefficients
should be extracted.}

\item{ic}{a character specifying the information criterion used for model selection.
Available options are \code{"AIC"}, \code{"BIC"} (default) and \code{"HQ"}.}

\item{select}{a character specifying how the best country model is selected.
If \code{"order"} (default), the country model with the overall minimum value of the
specified information criterion per country is selected. If \code{"rank"}, the
function selects the model, after which the selected information crition increases
for the first time.}
}
\value{
Data frame.
}
\description{
Calculates statistics for coefficient draws of contemporaneous foreign variables
and their domestic counterparts for the country models of a Bayesian GVAR model.
}
