% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_specifications.R
\name{create_specifications}
\alias{create_specifications}
\title{Create Country Model Specifications}
\usage{
create_specifications(
  country_data,
  global_data = NULL,
  countries = NULL,
  variables = NULL,
  type = "VAR",
  p_domestic = 2,
  p_foreign = 2,
  s = 2,
  r = NULL
)
}
\arguments{
\item{country_data}{a named list of time-series objects of country-specific data.}

\item{global_data}{a named time-series object of global data.}

\item{countries}{a character vector of country names, which should enter the GVAR model.
The names must correspond to the country names in \code{country_data}.}

\item{variables}{a character vector specifying the considered variables. If \code{NULL},
all available variables will be used.}

\item{type}{a character specifying if the country models are estimated as vector
autoregressive "VAR" (default) or error correction "VEC" models.}

\item{p_domestic}{an integer or integer vector of the lag order of domestic variables in the VAR. See also 'Details'.}

\item{p_foreign}{an integer or integer vector of the lag order of foreign variables in the VAR. See also 'Details'.}

\item{s}{an integer or integer vector of the lag order of global variables in the VAR. See also 'Details'.}

\item{r}{an integer or vector for the cointegration rank. See also 'Details'.}
}
\value{
The function produces a list of preliminary model specifications for each country,
which consists of the following elements:
\item{domestic}{a list of domestic variable names and their lag order.}
\item{foreign}{a list of foreign variable names and their lag order.}
\item{global}{a list of global variable names and their lag order (optional).}
\item{cointegration}{if \code{type = "VEC"}, a list of rank of the country model.}
\item{type}{a character specifying the type of the model, either "VAR" or "VEC".}
}
\description{
Produces a list of model specifications for every country model in a GVAR model.
}
\details{
The function assists in setting up a framework for the estimation of
the country-specific models of a GVAR model.

\code{p_domestic}, \code{p_star} and \code{s} refer to the lag order of the VAR formulation. If \code{type = "VEC"},
the function \code{\link{create_models}} will automatically reduce each lag order by 1. If a vector is provided,
\code{\link{create_models}} will produce a distinct model for all possible specification.

The argument \code{rank} is only used when \code{type = "VEC"}. Otherwise, the rank is set to zero.
If a vector is provided, \code{\link{create_models}} will produce a distinct model for all possible specifications.
}
\examples{
data("gvar2016")

country_data <- gvar2016$country_data
global_data <- gvar2016$global_data
region_weights <- gvar2016$region_weights
weight_data <- gvar2016$weight_data

# Generate weight matrices as 2 year, rolling window averages
gvar_weights <- create_weights(weight_data = weight_data, period = 2,
                               country_data = country_data)

# Create an object with country model specifications
model_specs <- create_specifications(country_data = country_data,
                                     global_data = global_data,
                                     variables = c("y", "Dp"),
                                     p_domestic = 2,
                                     p_foreign = 2,
                                     type = "VAR")

}
