% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctryvec_to_ctryvar.R
\name{ctryvec_to_ctryvar}
\alias{ctryvec_to_ctryvar}
\title{Transform the BVEC Submodels of a BGVAR model to a BVAR Submodel in Levels}
\usage{
ctryvec_to_ctryvar(object)
}
\arguments{
\item{object}{an object of class \code{"bgvecest"}.}
}
\value{
An object of class \code{"bgvarest"}.
}
\description{
An object of class \code{"bgvecest"} is transformed to a VAR in level representation.
}
\references{
Lütkepohl, H. (2006). \emph{New introduction to multiple time series analysis} (2nd ed.). Berlin: Springer.
}
