% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_methods.R
\name{diff.submodeldata}
\alias{diff.submodeldata}
\title{Differences of Variables}
\usage{
\method{diff}{submodeldata}(data, variables = NULL, multi = NULL)
}
\arguments{
\item{data}{an object of class 'submodeldata'.}

\item{variables}{a character vector of variables that should be differenced, if
they appear in a time-series object. If \code{NULL} (default), all variables are differenced.}

\item{multi}{optional. Numeric by which the differenced series should be multiplicated.}
}
\value{
A differenced time-series object or a list of differenced time-series objects.
}
\description{
Produces the first difference of variables in an object of class 'submodeldata'.
}
\examples{
# Load data
data("gvar2019")
submodel_data <- gvar2019$submodel_data

# Take first difference of the variables "y" and "Dp" across all
# elements of object "submodel_data" and multiply them by 100
submodel_data <- diff(submodel_data, variables = c("y", "Dp"), multi = 100)

}
