% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_posterior.gvecsubmodels.R
\name{draw_posterior.gvecsubmodels}
\alias{draw_posterior.gvecsubmodels}
\title{Posterior simulation}
\usage{
\method{draw_posterior}{gvecsubmodels}(object, ..., FUN = NULL, mc.cores = NULL, ctry = NULL)
}
\arguments{
\item{object}{a list of data and model specifications for country-specific
models, which should be passed on to function \code{FUN}. Usually, the
output of a call to \code{\link{create_models}} in combination with \code{\link{add_priors}}.}

\item{...}{further arguments passed to or from other methods.}

\item{FUN}{the function to be applied to each country model in argument \code{object}.
If \code{NULL} (default), the internal functions \code{\link{countryvarx}} is used for
VAR models and \code{\link{countryvecx}} for VEC models.}

\item{mc.cores}{the number of cores to use, i.e. at most how many child
processes will be run simultaneously. The option is initialized from
environment variable MC_CORES if set. Must be at least one, and
parallelization requires at least two cores.}

\item{ctry}{character vector specifying for which countries posterior distributions
should be drawn. If \code{NULL} (default), draws are generated for all country models.}
}
\value{
An object of class \code{bgvecest}, which contains a list of data,
model specifications, priors, coefficient draws and information
criteria for each estimated country model.
}
\description{
Estimates the country models of a Bayesian GVAR model.
}
