% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gvar.R
\name{estimate_gvar}
\alias{estimate_gvar}
\title{Fitter Function for GVAR Models}
\usage{
estimate_gvar(
  object,
  iterations = 50000,
  burnin = 5000,
  thin = 10,
  mc.cores = NULL
)
}
\arguments{
\item{object}{a list of data and model specifications for country-specific
models and information on the global VAR model as produced by the
\code{country_models} function.}

\item{iterations}{an integer of MCMC draws including burn-in (defaults
to 50000).}

\item{burnin}{an integer of MCMC draws used to initialize the sampler
(defaults to 5000). These draws do not enter the computation of posterior
moments, forecasts etc.}

\item{thin}{an integer specifying the thinning factor for the MCMC output.
Defaults to 10, which means that the forecast sequences contain only every
tenth draw of the original sequence. Set \code{thin = 1} to obtain the
full MCMC sequence.}

\item{mc.cores}{the number of cores to use, i.e. at most how many child
processes will be run simultaneously. The option is initialized from
environment variable MC_CORES if set. Must be at least one, and
parallelization requires at least two cores.}
}
\value{
An object of class \code{bgvarest} , which contains a list of data,
model specifications, priors, (thinned) coefficient draws and information
criteria for each estimated country model.
}
\description{
Estimates the country models of a Bayesian GVAR model.
}
\examples{
data("gvar2016")

country_data <- gvar2016$country_data
global_data <- gvar2016$global_data
region_weights <- gvar2016$region_weights
weight_data <- gvar2016$weight_data

# Take first difference of all variables y and Dp
country_data <- diff_variables(country_data, variables = c("y", "Dp", "r"))

# Generate EA area region with 2 year, rolling window weights
ea <- c("AT", "BE", "DE", "ES", "FI", "FR", "IT", "NL")
temp <- create_regions(country_data = country_data,
                       regions = list("EA" = ea),
                       period = 2,
                       region_weights = region_weights,
                       weight_data = weight_data)
country_data <- temp$country_data
weight_data <- temp$weight_data

# Generate weight matrices as 2 year rolling window averages
gvar_weights <- create_weights(weight_data = weight_data, period = 2,
                               country_data = country_data)

# Create an object with country model specifications
model_specs <- create_specifications(country_data = country_data,
                                     global_data = global_data,
                                     variables = c("y", "Dp", "r"),
                                     countries = c("EA", "US", "JP", "CA", "GB"),
                                     p_domestic = 1,
                                     p_foreign = 1,
                                     type = "VAR")

# Create estimable objects
object <- create_models(country_data = country_data,
                        gvar_weights = gvar_weights,
                        model_specs = model_specs)

# Add priors
object <- add_priors(object)

# Estimate GVAR model
gvar_est <- estimate_gvar(object, iterations = 50, burnin = 10, thin = 4)
# Note that the number of iterations and burnin should be much higher.

}
