% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_specifications_methods.R
\name{get_model_specifications.varxsubmodel}
\alias{get_model_specifications.varxsubmodel}
\title{Get Model Specifications}
\usage{
\method{get_model_specifications}{varxsubmodel}(object, ...)
}
\arguments{
\item{object}{an object of class \code{'varxsubmodel'}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A one-row data frame with the specifications of the sub-model: the
number of endogenous variables and their lag order, the number of weakly
exogenous and of global variables and their lag orders, the number of
deterministic terms, the number of observations if the object
contains data, and the variable selection algorithm.
}
\description{
Obtains the specifications of a sub-model of a global model.
}
