% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weight_matrix.R
\name{get_weight_matrix}
\alias{get_weight_matrix}
\title{Get Weight Matrix}
\usage{
get_weight_matrix(object, submodel)
}
\arguments{
\item{object}{an object of class 'gvarmodel'.}
}
\value{
A data frame.
}
\description{
Obtains the weight matrix of a sub-model.
}
