% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvec_to_gvar.R
\name{gvec_to_gvar}
\alias{gvec_to_gvar}
\title{Transform a Global VEC Model to a Global VAR Model}
\usage{
gvec_to_gvar(object)
}
\arguments{
\item{object}{an object of class 'gvecmodel' with estimated sub-models.}
}
\value{
An object of class 'gvarmodel'.
}
\description{
Rewrites every estimated sub-model of a global model in levels, so that the
global model can be solved.
}
\details{
\code{\link{submodels_to_gvar}} stacks sub-models in levels. A GVEC model
therefore has to be rewritten before it can be solved, which this function
does for all of its sub-models at once, using
\code{\link{vec_to_var.vecxsubmodel}} on each of them.

The global data, the variable index and the weight matrices are left
untouched: they describe the units of the model and their weights, neither of
which the transformation changes.
}
\examples{

# Load data
data("dees2007")
submodel_data <- dees2007[["submodel_data"]]
global_data <- dees2007[["global_data"]]

# Limit number of sub-models
submodel_data <- select_list_elements(submodel_data, c("US", "JP", "CA"))

# Set up the model
object <- create_gvecmodel(submodel_data = submodel_data,
                           global_data = global_data)
object <- add_weight_matrices(object = object,
                              submodel_data = submodel_data,
                              period = 1999:2001)
object <- add_submodels(object,
                        endogen = c("y", "Dp"), p_endogen = 1,
                        exogen = c("y", "Dp"), p_exogen = 1,
                        const = "unrestricted", trend = "restricted", r = 1,
                        iterations = 20, burnin = 10)
# Number of iterations and burn-in should be much higher.

object <- align_model_obs(object)
object <- add_priors(object,
                     coef = list(v_i = 0),
                     coint = list(v_i = 0, p_tau_i = 1),
                     sigma = list(df = 3, scale = 0.0001))
object <- add_initial_values(object)
object <- add_posterior_coefficients(object)

# Rewrite in levels and solve
object <- gvec_to_gvar(object)
gvar <- submodels_to_gvar(object)

}
