% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bgvarirf.R
\name{plot.bgvarirf}
\alias{plot.bgvarirf}
\title{Plotting Impulse Responses of Bayesian Vector Autoregression}
\usage{
\method{plot}{bgvarirf}(x, ...)
}
\arguments{
\item{x}{an object of class "bvarirf", usually, a result of a call to \code{\link{irf}}.}

\item{...}{further graphical parameters.}
}
\description{
A plot function for objects of class "bvarirf".
}
