% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minnesota_prior.vecxsubmodel.R
\name{minnesota_prior.vecxsubmodel}
\alias{minnesota_prior.vecxsubmodel}
\title{Minnesota Prior}
\usage{
\method{minnesota_prior}{vecxsubmodel}(
  object,
  kappa1 = 2,
  kappa2 = 0.5,
  kappa3 = NULL,
  kappa4 = 5,
  max_var = NULL,
  sigma = "AR"
)
}
\arguments{
\item{object}{an object of class 'vecxsubmodel', usually, a result of a call
to \code{\link{create_vecxsubmodel}}.}

\item{kappa1}{a numeric specifying the prior variance of coefficients that correspond to
own lags of endogenous variables. See 'Details'.}

\item{kappa2}{a numeric specifying the size of the prior variance of endogenous
variables, which do not correspond to own lags. See 'Details'.}

\item{kappa3}{a numeric specifying the size of the prior variance of non-deterministic exogenous
variables. Default is \code{NULL}, which indicates that the formula
for the calculation of the prior variance of deterministic terms is used
for all exogenous variables. See 'Details'.}

\item{kappa4}{a numeric specifying the size of the prior variance of deterministic
terms. See 'Details'.}

\item{max_var}{a positive numeric specifying the maximum prior variance that is allowed for
coefficients of non-deterministic variables. If \code{NULL} (default), the prior variances are not limited.}

\item{sigma}{either \code{"AR"} (default) or \code{"VAR"} indicating that the variances of the endogenous
variables \eqn{\sigma^2} are calculated based on a univariate AR regression or a least squares estimate of
the VAR form, respectively. In both cases all deterministic variables are used in the regressions,
if they appear in the model.}
}
\value{
A list containing a matrix of prior means and the precision matrix of the cofficients and the
inverse variance-covariance matrix of the error term, which was obtained by an LS estimation.
}
\description{
Calculates the Minnesota prior for a VECX sub-model.
}
\details{
The function calculates the Minnesota prior in the same manner as for
a traditional VAR model. For the endogenous variable
\eqn{i} the prior variance of the \eqn{l}th lag of regressor \eqn{j} is obtained as
\deqn{ \frac{\kappa_{1}}{l^2} \textrm{ for own lags of endogenous variables,}} 
\deqn{ \frac{\kappa_{1} \kappa_{2}}{l^2} \frac{\sigma_{i}^2}{\sigma_{j}^2} \textrm{ for endogenous variables other than own lags,}}
\deqn{ \frac{\kappa_{1} \kappa_{3}}{(l+1)^2} \frac{\sigma_{i}^2}{\sigma_{j}^2} \textrm{ for weakly exogenous and global variables,}}
\deqn{ \kappa_{1} \kappa_{4} \sigma_{i}^2 \textrm{ for deterministic terms,}}
where \eqn{\sigma_{i}} is the residual standard deviation of variable \eqn{i} of an unrestricted
LS estimate. For exogenous variables \eqn{\sigma_{i}} is the sample standard deviation.
In case structural parameters are estimated, the formula
\eqn{\kappa_{1} \kappa_{2} \frac{\sigma_{i}^2}{\sigma_{j}^2}} is used.
If \eqn{kappa_{3}} is not provided, prior variances are calculated in the same way as for
deterministic terms.
If the model does not contain exogenous variables, argument \code{kappa3} will be ignored.

The function only provides priors for the non-cointegration part of the model. The
residual standard errors \eqn{\sigma_i} are based on an unrestricted LS regression of the
endogenous variables on the error correction term and the non-cointegration regressors.
}
\references{
Chan, J., Koop, G., Poirier, D. J., & Tobias, J. L. (2020). \emph{Bayesian Econometric Methods}
(2nd ed.). Cambridge: University Press.

Lütkepohl, H. (2006). \emph{New introduction to multiple time series analysis} (2nd ed.). Berlin: Springer.
}
