% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bgvarest.R
\name{plot.bgvarest}
\alias{plot.bgvarest}
\title{Plotting Draws of a Bayesian Global VAR Models}
\usage{
\method{plot}{bgvarest}(x, ci = 0.95, type = "hist", ctry = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{"bgvarest"}, usually, a result of a call to \code{\link{draw_posterior}}.}

\item{ci}{interval used to calculate credible bands for time-varying parameters.}

\item{type}{either \code{"hist"} (default) for histograms, \code{"trace"} for a trace plot,
or \code{"boxplot"} for a boxplot. Only used for parameter draws of constant coefficients.}

\item{ctry}{character. Name of the element in argument \code{x}, for which posterior draws
should be plotted. If \code{NULL} (default), all submodels are used.}

\item{...}{further graphical parameters.}
}
\description{
A plot function for objects of class \code{"bgvarest"}.
}
