% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bgvarestirf.R
\name{plot.bgvarestirf}
\alias{plot.bgvarestirf}
\title{Plotting Draws of a Bayesian Global VAR Models}
\usage{
\method{plot}{bgvarestirf}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"bgvarestirf"}, usually, a result of a call to \code{\link{irf.bgvarest}}.}

\item{...}{further graphical parameters.}
}
\description{
A plot function for objects of class \code{"bgvarest"}.
}
