% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bgvarprd.R
\name{plot.bgvarprd}
\alias{plot.bgvarprd}
\title{Plotting Forecasts of BVAR Models}
\usage{
\method{plot}{bgvarprd}(x, variable = NULL, ...)
}
\arguments{
\item{x}{an object of class "bvarprd", usually, a result of a call to \code{\link{predict.bgvar}}.}

\item{variable}{a character vector containing a country name as its first and a variable name
as its second element. If \code{NULL} (default) all series are plotted.}

\item{...}{further graphical parameters.}
}
\description{
A plot function for objects of class "bvarprd".
}
