% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ctryvartest.R
\name{plot.ctryvartest}
\alias{plot.ctryvartest}
\title{Plotting Submodel Test Statistics}
\usage{
\method{plot}{ctryvartest}(x, ctry = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{"ctryvarest"}, usually, a result of a call
to \code{\link{draw_posterior}}.}

\item{ctry}{character vector (optional). Names of the countries, for which test
statistics should be visualized.}

\item{...}{further graphical parameters.}
}
\description{
A plot function for objects of class \code{"ctryvartest"} for visual inspection
of posterior draws.
}
