% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ctryvecest.R
\name{plot.ctryvecest}
\alias{plot.ctryvecest}
\title{Plotting Draws of a VECX Submodel of a GVAR Model}
\usage{
\method{plot}{ctryvecest}(x, ci = 0.95, type = "hist", variables = "all", ctry = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{"ctryvecest"}, usually, a result of a call to \code{\link{draw_posterior}}.}

\item{ci}{interval used to calculate credible bands for time-varying parameters.}

\item{type}{either \code{"hist"} (default) for histograms, \code{"trace"} for a trace plot,
or \code{"boxplot"} for a boxplot. Only used for parameter draws of constant coefficients.}

\item{variables}{character vector of variables that should be plotted. Default is \code{"all"}.
Other options are \code{"domestic"}, \code{"foreign"}, \code{"global"}, \code{"deterministic"}
and \code{"sigma"}.}

\item{ctry}{character (optional). Name of the country, which will be shown in the title of the output.}

\item{...}{further graphical parameters.}
}
\description{
A plot function for objects of class \code{"ctryvecest"} for visual inspection
of posterior draws.
}
