% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gvarmodel.R
\name{print.gvarmodel}
\alias{print.gvarmodel}
\title{Printing Model Information}
\usage{
\method{print}{gvarmodel}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class 'gvarmodel'.}

\item{digits}{the number of significant digits to use when printing.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The object that was printed, invisibly. A table of the variables of
every sub-model is printed as a side effect, with a mark for each variable
that the sub-model contains.
}
\description{
print method for objects of class 'gvarmodel'.
}
