% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.submodeldata.R
\name{print.submodeldata}
\alias{print.submodeldata}
\title{Printing the Data of Sub-Models}
\usage{
\method{print}{submodeldata}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class 'submodeldata'.}

\item{digits}{the number of significant digits to use when printing.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The printed table, invisibly: one row per sub-model and one column
per variable, with a mark for each variable the sub-model provides.
}
\description{
print method for objects of class 'submodeldata'.
}
