% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_error_correction.R
\name{rescale_error_correction.vecxsubmodel}
\alias{rescale_error_correction.vecxsubmodel}
\title{Rescale Error Correction}
\usage{
\method{rescale_error_correction}{vecxsubmodel}(object, ...)
}
\arguments{
\item{object}{object of class 'vecxsubmodel'.}

\item{...}{arguments passed forward to method.}
}
\value{
An object of class 'vecxsubmodel'.
}
\description{
Rescales the series in the error correction series in an object of class
'vecxsubmodel'.
}
\details{
The function transforms element \code{object$data$train$w} and
the posterior draws of \eqn{alpha} and \eqn{beta} to the original scale
of the input data, based on the scaling factors that are stored in attribute
\code{"scale"} of \code{object$data$train$w}.
}
