% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_list_elements.R
\name{select_list_elements}
\alias{select_list_elements}
\title{Select List Elements}
\usage{
select_list_elements(object, submodel)
}
\arguments{
\item{object}{a named list whose elements should be filtered.}

\item{submodel}{character vector of the elements in \code{object}, which
should remain in the output.}
}
\description{
Limits the elements of a named list to the specified entities, preserving
all class information.
}
