% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_submodels.R
\name{select_submodels}
\alias{select_submodels}
\title{Submodel Selection}
\usage{
select_submodels(object, models)
}
\arguments{
\item{object}{a list containing the posterior draws of the submodels of a
GVAR model. Usually, the result of a call to \code{\link[bvartools]{add_posterior_coefficients}}.}

\item{models}{a vector of integers with the position of the submodels, which
should be used in the global model.}
}
\description{
Creates a list of submodels of a GVAR model, which are used to solve the
global model.
}
