% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_gvar.R
\name{solve_gvar}
\alias{solve_gvar}
\title{Solve a GVAR Model}
\usage{
solve_gvar(object, ic = "BIC", t = NULL, thin = 1, select = "order")
}
\arguments{
\item{object}{a list containing the results of the country estimates, usually, the
result of a call to \code{\link{estimate_gvar}}.}

\item{ic}{a character specifying the information criterion used for model selection.
Available options are "AIC", "BIC" (default) and "HQ".}

\item{t}{an integer of the time index for which the GVAR should be solved. Only used
when time varying weights or parameters are used.}

\item{thin}{an integer specifying the thinning factor for the MCMC output.
Defaults to 1 to obtain the full MCMC sequence.}

\item{select}{a character specifying how the best country model is selected.
If \code{"order"} (default), the country model with the overall minimum value of the
specified information criterion per country is selected. If \code{"rank"}, the
function selects the model, after which the selected information crition increases
for the first time.}
}
\value{
An object of class \code{"bgvar"}.
}
\description{
Combines the country model results to a global VAR model and solves it.
}
\details{
If multiple country models were estimated with \code{\link{estimate_gvar}},
the function will choose the model, which maximises the specified information
criterion specified by \code{ic}.
}
\examples{
data("gvar2016")

country_data <- gvar2016$country_data
global_data <- gvar2016$global_data
region_weights <- gvar2016$region_weights
weight_data <- gvar2016$weight_data

# Take first difference of all variables y and Dp
country_data <- diff_variables(country_data, variables = c("y", "Dp", "r"))

# Generate EA area region with 2 year rolling window weights
ea <- c("AT", "BE", "DE", "ES", "FI", "FR", "IT", "NL")
temp <- create_regions(country_data = country_data,
                       regions = list("EA" = ea),
                       period = 2,
                       region_weights = region_weights,
                       weight_data = weight_data)
country_data <- temp$country_data
weight_data <- temp$weight_data

# Generate weight matrices as 2 year rolling window averages
gvar_weights <- create_weights(weight_data = weight_data, period = 2,
                               country_data = country_data)

# Create an object with country model specifications
model_specs <- create_specifications(country_data = country_data,
                                     global_data = global_data,
                                     variables = c("y", "Dp", "r"),
                                     countries = c("EA", "US", "JP", "CA", "MX", "GB"),
                                     p_domestic = 1,
                                     p_foreign = 1,
                                     type = "VAR")

# Create estimable objects
object <- create_models(country_data = country_data,
                        gvar_weights = gvar_weights,
                        model_specs = model_specs)

# Add priors
object <- add_priors(object)

# Estimate GVAR model
gvar_est <- estimate_gvar(object, iterations = 100, burnin = 10, thin = 2)
# Note that the number of iterations and burnin should be much higher, e.g., 30000.

# Solve GVAR
gvar_solved <- solve_gvar(gvar_est)

}
