% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodel_test_statistics.R
\name{submodel_test_satistics}
\alias{submodel_test_satistics}
\title{Test Statistics for Model Selection of VARX Submodels}
\usage{
submodel_test_satistics(object, ...)
}
\arguments{
\item{object}{an object of class \code{"bgvarest"}, usually, a result of a call to
\code{\link{draw_posterior}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A data frame.
}
\description{
Calculates test stistics for the the selection of country-specific VARX models
of a GVAR model.
}
\details{
The log-likelihood for the calculation of the information criteria is obtained by
\deqn{LL = \frac{1}{R} \sum_{i = 1}^{R} \left( \sum_{t = 1}^{T} -\frac{K^{dom}}{2} \ln 2\pi - \frac{1}{2} \ln |\Sigma_t^{(i)}| -\frac{1}{2} (u_t^{{(i)}\prime} (\Sigma_t^{(i)})^{-1} u_t^{(i)} \right)},
where \eqn{u_t = y_t - \mu_t}. The Akaike, Bayesian and Hannan–Quinn (HQ) information criteria are calculated as
\deqn{AIC = 2 (K^{d}p^{d} + K^{f}p^{f} + Ms + N) - 2 LL},
\deqn{BIC = (K^{d}p^{d} + K^{f}p^{f} + Ms + N) ln(T) - 2 LL} and 
\deqn{HQ = 2  (K^{d}p^{d} + K^{f}p^{f} + Ms + N) ln(ln(T)) - 2 LL}, respectively,
where \eqn{K^{d}} is the number of endogenous domestic variables, \eqn{p^{d}} the number of lags of endogenous domestic variables,
\eqn{K^{f}} is the number of foreign variables, \eqn{p^{f}} the number of lags of foreign variables,
\eqn{M} the number of global variables, \eqn{s} the number of lags of global variables,
\eqn{N} the number of deterministic terms and \eqn{T} the number of observations.
}
