% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bgvarest.R, R/print.summary.bgvarest.R
\name{summary.bgvarest}
\alias{summary.bgvarest}
\alias{print.summary.bgvarest}
\title{Summarising Country-Specific VARX Submodels of a GVAR Model}
\usage{
\method{summary}{bgvarest}(object, ci = 0.95, period = NULL, ctry = NULL, ...)

\method{print}{summary.bgvarest}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{an object of class \code{"bgvarest"}, usually, a result of a call to
\code{\link{draw_posterior}}.}

\item{ci}{a numeric between 0 and 1 specifying the probability of the credible band.
Defaults to 0.95.}

\item{period}{integer. Index of the period, for which the summary statistics should be generated.
Only used for TVP or SV models. Default is \code{NULL}, so that the posterior draws of the last time period
are used.}

\item{ctry}{character. Name of the element in argument \code{object}, for which summary
statistics should be obtained. If \code{NULL} (default), all country submodels are used.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class \code{"summary.bgvarest"}, usually, a result of a call to
\code{\link{summary.bgvarest}}.}

\item{digits}{the number of significant digits to use when printing.}
}
\value{
\code{summary.bgvarest} returns a list of class \code{"summary.bgvarest"},
which contains summary statistics of country-specific submodels of class
\code{\link{summary.ctryvarest}}.
}
\description{
summary method for class \code{"bgvarest"}.
}
