% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ctryvarest.R,
%   R/print.summary.ctryvarest.R
\name{summary.ctryvarest}
\alias{summary.ctryvarest}
\alias{print.summary.ctryvarest}
\title{Summarising Country-Specific VARX Submodels of a GVAR Model}
\usage{
\method{summary}{ctryvarest}(object, ci = 0.95, period = NULL, ...)

\method{print}{summary.ctryvarest}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{an object of class \code{"ctryvarest"}, usually, a result of a call to \code{\link{draw_posterior}}.}

\item{ci}{a numeric between 0 and 1 specifying the probability of the credible band.
Defaults to 0.95.}

\item{period}{integer. Index of the period, for which the summary statistics should be generated.
Only used for TVP or SV models. Default is \code{NULL}, so that the posterior draws of the last time period
are used.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an object of class \code{"summary.ctryvarest"}, usually, a result of a call to
\code{\link{summary.ctryvarest}}.}

\item{digits}{the number of significant digits to use when printing.}
}
\value{
\code{summary.ctryvarest} returns a list of class \code{"summary.ctryvarest"},
which contains the following components:
\item{coefficients}{A list of various summary statistics of the posterior
draws of the VAR coefficients.}
\item{sigma}{A list of various summary statistics of the posterior
draws of the variance-covariance matrix.}
\item{model}{a list containing information on the model specifications.}
}
\description{
summary method for class \code{"ctryvarest"}.
}
