% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin_methods.R
\name{thin.varxsubmodelest}
\alias{thin.varxsubmodelest}
\title{Thinning Posterior Draws}
\usage{
\method{thin}{varxsubmodelest}(x, thin = 10, ...)
}
\arguments{
\item{x}{an object of class 'varxsubmodelest'.}

\item{thin}{an integer specifying the thinning interval between successive values of posterior draws.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object of class 'varxsubmodelest'.
}
\description{
Thins the MCMC posterior draws in an object of class 'varxsubmodelest'.
}
