% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_expanding_window.varxsubmodel.R
\name{use_expanding_window.varxsubmodel}
\alias{use_expanding_window.varxsubmodel}
\title{Expanding Window Estimation}
\usage{
\method{use_expanding_window}{varxsubmodel}(object, start, ...)
}
\arguments{
\item{object}{a list of model specifications. Usually, the output of a call to 
a model creation function in combination with \code{\link[bvartoosl]{add_priors}} and
\code{\link[bvartools]{add_initial_values}}.}

\item{start}{the start period of the prediction of the first iteration of the
expanding window approach.}

\item{...}{arguments passed forward to method.}
}
\description{
Creates objects for expanding window posterior simulation.
}
