% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_hdf5.gvecmodel.R
\name{write_to_hdf5.gvecmodel}
\alias{write_to_hdf5.gvecmodel}
\title{Export to HDF5 File}
\usage{
\method{write_to_hdf5}{gvecmodel}(object, folder, ...)
}
\arguments{
\item{object}{list of class 'gvecmodel'.}

\item{folder}{path to the folder where the content of argument
\code{object} should be saved.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Exports the content of an object of class 'gvecmodel' to a dedicated folder.
}
