% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_initial_values.gvarmodel.R
\name{add_initial_values.gvarmodel}
\alias{add_initial_values.gvarmodel}
\title{Add Initial Values of an MCMC Chain}
\usage{
\method{add_initial_values}{gvarmodel}(object, ...)
}
\arguments{
\item{object}{an object of class 'gvarmodel', usually, the output of a call to
\code{\link{create_gvarmodel}} in combination with \code{\link{add_submodels}}
and \code{\link[bvartools]{add_initial_values}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list of class 'gvarmodel'.
}
\description{
Adds initial values to a list of models by passing each element to
the respective method.
}
\examples{

# Load data
data("gvar2019")
global_data <- gvar2019[["global_data"]]
submodel_data <- gvar2019[["submodel_data"]]

# Limit number of sub-models
submodel_data <- select_list_elements(submodel_data, c("AT", "DE", "US"))

# Create global model
object <- create_gvarmodel(submodel_data = submodel_data,
                           global_data = global_data)

# Generate and add weight matrices
object <- add_weight_matrices(object = object,
                              submodel_data = submodel_data,
                              period = 2013:2016)

object <- add_submodels(object,
                        p_endogen = 1,
                        p_exogen = 2,
                        global = "poil",
                        s = 2,
                        error = "wishart",
                        iterations = 10,
                        burnin = 10)
# Number of iterations and burn-in should be much higher.

# Add priors
object <- add_priors(object,
                     coef = list(v_i = 0),
                     sigma = list(df = 3, scale = 0.0001))
                     
# Add initial values
object <- add_initial_values(object)

}
