% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_initial_values.gvecmodel.R
\name{add_initial_values.gvecmodel}
\alias{add_initial_values.gvecmodel}
\title{Add Initial Values of an MCMC Chain}
\usage{
\method{add_initial_values}{gvecmodel}(object, ...)
}
\arguments{
\item{object}{an object of class 'gvecmodel', usually, the output of a call to
\code{\link{create_gvecmodel}} in combination with \code{\link{add_submodels}}
and \code{\link[bvartools]{add_initial_values}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list of class 'gvecmodel'.
}
\description{
Adds initial values to a list of models by passing each element to
the respective method.
}
