% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_priors.gvecmodel.R
\name{add_priors.gvecmodel}
\alias{add_priors.gvecmodel}
\title{Add Priors to Bayesian Models}
\usage{
\method{add_priors}{gvecmodel}(object, ...)
}
\arguments{
\item{object}{a list of models, usually, the output of a call to
\code{\link{create_gvecmodel}} in combination with \code{\link{add_submodels}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list of class 'gvecmodel'.
}
\description{
Adds prior specifications to a list of models by passing each element to
the respective method.
}
\examples{

# Load data
data("dees2007")
submodel_data <- dees2007[["submodel_data"]]
global_data <- dees2007[["global_data"]]

# Limit number of sub-models
submodel_data <- select_list_elements(submodel_data, c("EA", "US"))

# Create empty model
object <- create_gvecmodel(submodel_data = submodel_data,
                           global_data = global_data)

# Add weight matrices
object <- add_weight_matrices(object = object,
                              submodel_data = submodel_data,
                              period = 1999:2001)

# Create sub-models
object <- add_submodels(object,
                        p_endogen = 1, p_exogen = 1,
                        global = "poil",
                        s = 1,
                        r = 1,
                        error = "wishart",
                        iterations = 10,
                        burnin = 10)
# Number of iterations and burn-in should be much higher.

# Add priors
object <- add_priors(object,
                     coef = list(v_i = 0),
                     coint = list(v_i = 0, p_tau_i = 1),
                     sigma = list(df = 3, scale = 0.0001))

}
