% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_priors.vecxsubmodel.R
\name{add_priors.vecxsubmodel}
\alias{add_priors.vecxsubmodel}
\title{Add Priors to Bayesian Models}
\usage{
\method{add_priors}{vecxsubmodel}(object, coef, coint, sigma, varsel = NULL, ...)
}
\arguments{
\item{object}{a list of class 'vecxsubmodel'.}

\item{coef}{a named list of prior specifications for coefficients that do not 
determine the cointegration space. For the default specification all prior means are set to zero
and the diagonal elements of the inverse prior variance-covariance matrix are set to 1
for coefficients corresponding to non-deterministic terms. For deterministic coefficients the prior
variances are set to 10 by \code{v_i_det = 0.1}. The variances need to be specified as precisions,
i.e. as inverses of the variances. For further specifications such as the Minnesota prior see 'Details'.}

\item{coint}{a named list of prior specifications for coefficients determining the
cointegration space of VEC models. See 'Details'.}

\item{sigma}{a named list of prior specifications for the error variance-covariance matrix.
For the default specification of an inverse Wishart distribution the prior degrees
of freedom are set to the number of endogenous variables plus the rank of
the cointegration matrix. The prior variance is to 1. See 'Details'.}

\item{varsel}{optional; a named list of prior specifications for the variable
selection algorithm. See 'Details'.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object of class 'vecxsubmodel'.
}
\description{
Adds prior specifications to a list of models, which was produced by
function \code{\link{create_vecxsubmodel}}.
}
\details{
The arguments of the function require named lists. Possible
specifications are described in the following. Note that it is important to specify the
priors in the correct list. Otherwise, the provided specification will be disregarded
and default values will be used.

Argument \code{coef} contains the following elements
\describe{
  \item{\code{v_i}}{a numeric specifying the prior precision of the coefficients. Default is 1.
  Will be ignored if \code{minnesota} is specified.}
  \item{\code{v_i_det}}{a numeric specifying the prior precision of coefficients
  corresponding to deterministic terms. Default is 0.1. Will be ignored if \code{minnesota} is specified.}
  \item{\code{const}}{a numeric or character specifying the prior mean of coefficients, which correspond
  to the intercept. If a numeric is provided, all prior means are set to this value.
  If \code{coef$const = "mean"}, the mean of the respective endogenous variable is used as prior mean.
  If \code{coef$const = "first"}, the first values of the respective endogenous variable is used as prior mean.}
  \item{\code{minnesota}}{a named list containing for parameters for the calculation of
  the Minnesota prior, where the element names must be \code{kappa1}, \code{kappa2}, \code{kappa3} and \code{kappa4}.
  For the endogenous variable \eqn{i} the prior variance of the \eqn{l}th lag of regressor \eqn{j} is obtained as
  \deqn{ \frac{\kappa_{1}}{l^2} \textrm{ for own lags of endogenous variables,}} 
  \deqn{ \frac{\kappa_{1} \kappa_{2}}{l^2} \frac{\sigma_{i}^2}{\sigma_{j}^2} \textrm{ for endogenous variables other than own lags,}}
  \deqn{ \frac{\kappa_{1} \kappa_{3}}{(l+1)^2} \frac{\sigma_{i}^2}{\sigma_{j}^2} \textrm{ for exogenous variables,}}
  \deqn{ \kappa_{1} \kappa_{4} \sigma_{i}^2 \textrm{ for deterministic terms,}}
  where \eqn{\sigma_{i}} is the residual standard deviation of variable \eqn{i} of an unrestricted
  LS estimate. For exogenous variables \eqn{\sigma_{i}} is the sample standard deviation.
  The function only provides priors for the non-cointegration part of the model. However,
  the residual standard errors \eqn{\sigma_i} are based on an unrestricted LS regression of the
  endogenous variables on the error correction term and the non-cointegration regressors.}
  \item{\code{max_var}}{a numeric specifying the maximum prior variance that is allowed for
  non-deterministic coefficients. This argument can be used to overwrite prior variances that are
  based on LS estimates, but should not exceed a certain threshold. This can
  be useful when using the BVS algorithm.}
  \item{\code{shape}}{a numeric specifying the prior shape parameter of the error term of the
  state equation. Only used for models with time varying parameters. Default is 3.}
  \item{\code{rate}}{a numeric specifying the prior rate parameter of the error term of the
  state equation. Only used for models with time varying parameters. Default is 0.0001.}
  \item{\code{rate_det}}{a numeric specifying the prior rate parameter of the error term of the
  state equation for coefficients, which correspond to deterministic terms.
  Only used for models with time varying parameters. Default is 0.01.}
}

Argument \code{coint} depends on whether the cointegration parameters are constant or time
varying, since the two are given different kinds of prior: a matric-variate prior on the
cointegration space in the first case, and a state equation in the second.
\describe{
  \item{\code{v_i}}{numeric between 0 and 1 specifying the shrinkage of the cointegration space prior.
  Required for sub-models with constant cointegration parameters and not used otherwise.}
  \item{\code{p_tau_i}}{numeric of the diagonal elements of the inverse prior matrix of
  the central location of the cointegration space \eqn{sp(\beta)}.
  Required for sub-models with constant cointegration parameters and not used otherwise.}
  \item{\code{rho}}{a numeric specifying the autocorrelation coefficient
  of the state equation of \eqn{\beta}. It must be smaller than 1.
  Required for sub-models with time varying cointegration parameters and not used otherwise.
  Note that in contrast to Koop et al. (2011) \eqn{\rho} is not drawn in the Gibbs sampler of
  this package yet.}
}
For a sub-model with time varying cointegration parameters the state equation is
\eqn{\beta_t = \rho \beta_{t-1} + \eta_t} with \eqn{\eta_t \sim N(0, I)}, and
the prior on the state before the sample is that equation's own stationary
distribution, \eqn{N(0, I / (1 - \rho^2))}. This is what makes the prior
proper, so a \eqn{\rho} close to one is intended and one further from it draws
a warning. The loadings carry the compensating scale: only the product
\eqn{\alpha \beta^{\prime}} is identified, so their prior variance is shrunk
by \eqn{1 - \rho^2}, leaving the product on the scale \code{coef$v_i} asks
for.

Variable selection is available as \code{"bvs"}. The loadings are never subject to it,
so the error correction term is always in the model, and one inclusion parameter is drawn
per coefficient rather than per coefficient and period. SSVS is not implemented for error
correction sub-models and is refused when the priors are added.

Argument \code{sigma} can contain the following elements:
\describe{
  \item{\code{df}}{an integer or character specifying the prior degrees of freedom of the error term. Only
  used, if the prior is inverse Wishart.
  Default is \code{"k"}, which indicates the amount of endogenous variables in the respective model.
  \code{"k + 3"} can be used to set the prior to the amount of endogenous variables plus 3. If an integer
  is provided, the degrees of freedom are set to this value in all models.
  In all cases the rank \eqn{r} of the cointegration matrix is automatically added.}
  \item{\code{scale}}{a numeric specifying the prior error variance of endogenous variables.
  Default is 1.}
  \item{\code{shape}}{a numeric or character specifying the prior shape parameter of the error term. Only
  used, if the prior is inverse gamma or if time varying volatilities are estimated.
  For models with constant volatility the default is \code{"k"}, which indicates the amount of endogenous
  variables of the model. \code{"k + 3"} can be used to set the prior to the amount of
  endogenous variables plus 3. If a numeric is provided, the shape parameters are set to this value in all
  models. For models with stochastic volatility this prior refers to the error variance of the state
  equation.}
  \item{\code{rate}}{a numeric specifying the prior rate parameter of the error term. Only used, if the
  prior is inverse gamma or if time varying volatilities are estimated. For models with stochastic
  volatility this prior refers to the error variance of the state equation.}
  \item{\code{mu}}{numeric of the prior mean of the initial state of the log-volatilities.
  Only used for models with time varying volatility.}
  \item{\code{v_i}}{numeric of the prior precision of the initial state of the log-volatilities.
  Only used for models with time varying volatility.}
  \item{\code{state_variance}}{numeric of the initial draw for the variance of the log-volatilities.
  Only used for models with time varying volatility.}
  \item{\code{offset}}{numeric of the constant, which is added before taking the log of the squared errors.
  Only used for models with time varying volatility.}
}
For structural models only a gamma prior or stochastic volatility specification is allowed.

Argument \code{varsel} can contain the following elements:
\describe{
  \item{\code{inprior}}{a numeric between 0 and 1 specifying the prior probability
  of a variable to be included in the model.}
  \item{\code{covar}}{logical indicating if the variable selection algorithm
  should also be applied to the error covariance matrix.}
  \item{\code{exclude_det}}{logical indicating if deterministic terms should
  be excluded from the variable selection algorithm.}
  \item{\code{minnesota}}{a numeric vector of length 4 containing parameters
  for the calculation of the Minnesota-like inclusion priors. See below.}
  \item{\code{tau}}{a numeric vector of two elements containing the prior standard errors
  of restricted variables (\eqn{\tau_0}) as its first element and unrestricted variables (\eqn{\tau_1})
  as its second. Only used for SSVS.}
  \item{\code{semiautomatic}}{an numeric vector of two elements containing the
  factors by which the standard errors associated with an unconstrained least squares
  estimate of the model are multiplied to obtain the prior standard errors
  of restricted (\eqn{\tau_0}) and unrestricted (\eqn{\tau_1}) variables, respectively.
  This is the semiautomatic approach described in George et al. (2008). Only used for SSVS.}
}
In the case of SSVS, either \code{tau} or \code{semiautomatic} must be specified.

If \code{varsel$minnesota} is specified, prior
inclusion probabilities are calculated in a Minnesota-like fashion as
\tabular{cl}{
\eqn{\frac{\kappa_1}{l}} \tab for own lags of endogenous variables, \cr
\eqn{\frac{\kappa_2}{l}} \tab for other endogenous variables, \cr
\eqn{\frac{\kappa_3}{1 + l}} \tab for exogenous variables, \cr
\eqn{\kappa_{4}} \tab for deterministic variables, 
}
for lag \eqn{l} with \eqn{\kappa_1}, \eqn{\kappa_2}, \eqn{\kappa_3},
\eqn{\kappa_4} as the first, second, third and forth element in
\code{varsel$minnesota}, respectively.
}
\examples{

# Load data
data("dees2007")
submodel_data <- dees2007[["submodel_data"]]
global_data <- dees2007[["global_data"]]

# Limit number of sub-models
submodel_data <- select_list_elements(submodel_data, c("EA", "US"))

# Create empty model
object <- create_gvecmodel(submodel_data = submodel_data,
                           global_data = global_data)

# Add weight matrices
object <- add_weight_matrices(object = object,
                              submodel_data = submodel_data,
                              period = 1999:2001)

# Create sub-models
object <- add_submodels(object,
                        p_endogen = 1, p_exogen = 1,
                        global = "poil",
                        s = 1,
                        r = 1,
                        error = "wishart",
                        iterations = 10,
                        burnin = 10)
# Number of iterations and burn-in should be much higher.

# Add priors
object <- add_priors(object[["submodels"]][["EA"]][[1]],
                     coef = list(v_i = 0),
                     coint = list(v_i = 0, p_tau_i = 1),
                     sigma = list(df = 3, scale = 0.0001))

}
\references{
Chan, J., Koop, G., Poirier, D. J., & Tobias J. L. (2019). \emph{Bayesian econometric methods}
(2nd ed.). Cambridge: Cambridge University Press.

George, E. I., Sun, D., & Ni, S. (2008). Bayesian stochastic search for VAR model
restrictions. \emph{Journal of Econometrics, 142}(1), 553--580.
\doi{10.1016/j.jeconom.2007.08.017}

Koop, G., León-González, R., & Strachan R. W. (2010). Efficient posterior
simulation for cointegrated models with priors on the cointegration space.
\emph{Econometric Reviews, 29}(2), 224--242.
\doi{10.1080/07474930903382208}

Koop, G., León-González, R., & Strachan R. W. (2011). Bayesian inference in
a time varying cointegration model. \emph{Journal of Econometrics, 165}(2), 210--220.
\doi{10.1016/j.jeconom.2011.07.007}

Korobilis, D. (2013). VAR forecasting using Bayesian variable selection.
\emph{Journal of Applied Econometrics, 28}(2), 204--230. \doi{10.1002/jae.1271}

Lütkepohl, H. (2006). \emph{New introduction to multiple time series analysis} (2nd ed.). Berlin: Springer.
}
