% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_weight_matrices.R
\name{add_weight_matrices}
\alias{add_weight_matrices}
\title{Add Weight Matrices
 
Generic function used to generate weight matrices and add them to a model object.}
\usage{
add_weight_matrices(object, ...)
}
\arguments{
\item{object}{an object of a class, for which a method should be called.}

\item{...}{arguments passed forward to method.}
}
\description{
Add Weight Matrices
 
Generic function used to generate weight matrices and add them to a model object.
}
