% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgvars-package.R
\docType{package}
\name{bgvars-package}
\alias{bgvars}
\alias{bgvars-package}
\title{bgvars: Bayesian Inference of Global Vector Autoregressive and Gloal Vector Error Correction Models}
\description{
Assists in the set-up and inference of Bayesian global vector autoregressive (BGVAR) and error correction (BGVEC) models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/franzmohr/bgvars}
  \item Report bugs at \url{https://github.com/franzmohr/bgvars/issues}
}

}
\author{
\strong{Maintainer}: Franz X. Mohr \email{franz.x.mohr@outlook.com} (0009-0003-8890-7781)

}
\keyword{internal}
