% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgvars.R
\docType{package}
\name{bgvars}
\alias{bgvars}
\alias{bgvars-package}
\title{bgvars: Bayesian Inference for Global Vector Autoregressive (GVAR) and Global Vector Error Correction (GVEC) Models}
\description{
R/C++ implementation of Bayesian global vector autoregressive and global vector error correction models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/franzmohr/bgvars}
  \item Report bugs at \url{https://github.com/franzmohr/bgvars/issues}
}

}
\author{
Franz X. Mohr
}
