% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgvars.R
\docType{package}
\name{bgvars}
\alias{bgvars}
\title{bgvars: Bayesian Inference for Global Vector Autoregressive Models}
\description{
R/C++ implementation of Bayesian global vector autoregressive models.
}
\author{
Franz X. Mohr
}
