% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_submodels.R
\name{combine_submodels}
\alias{combine_submodels}
\title{Generate a GVAR Model}
\usage{
combine_submodels(object, period = NULL, thin = 1)
}
\arguments{
\item{object}{a list containing the results of the country estimates, usually, the
result of a call to \code{\link{draw_posterior}}.}

\item{period}{an integer of the time index for which the GVAR should be solved. Only used
when time varying weights or parameters are used.}

\item{thin}{an integer specifying the thinning factor for the MCMC output.
Defaults to 1 to obtain the full MCMC sequence.}

\item{ic}{a character specifying the information criterion used for model selection.
Available options are "AIC", "BIC" (default) and "HQ".}

\item{select}{a character specifying how the best country model is selected.
If \code{"order"} (default), the country model with the overall minimum value of the
specified information criterion per country is selected. If \code{"rank"}, the
function selects the model, after which the selected information crition increases
for the first time.}
}
\value{
An object of class \code{"bgvar"}.
}
\description{
Combines the country model results to a global VAR model and solves it.
}
\examples{

# Load gvar2016 dataset
data("gvar2016")

# Data objects
country_data <- gvar2016[["country_data"]]
global_data <- gvar2016[["global_data"]]
region_weights <- gvar2016[["region_weights"]]
weight_data <- gvar2016[["weight_data"]]

# Obtain weights
weight_data <- create_weights(weight_data = weight_data, period = 3,
                              country_data = country_data)

# Generate specifications
model_specs <- create_specifications(country_data = country_data,
                                     global_data = global_data,
                                     countries = c("US", "JP", "CA", "NO", "GB"), 
                                     domestic = list(variables = c("y", "Dp", "r"), lags = 1),
                                     foreign = list(variables = c("y", "Dp", "r"), lags = 1),
                                     global = list(variables = c("poil"), lags = 1),
                                     deterministic = list(const = TRUE, trend = FALSE, seasonal = FALSE),
                                     iterations = 10,
                                     burnin = 10,
                                     thin = 1)
                                     
# Note that the number of iterations and burnin draws is usually much higher.

# Create list element for each model
country_models <- create_models(country_data = country_data,
                                weight_data = weight_data,
                                global_data = global_data,
                                model_specs = model_specs)
                                
# Add priors
country_models <- add_priors(country_models)

# Draw from posterior distribution
gvar_est <- draw_posterior(country_models, mc.cores = 1)

# Solve model
gvar <- combine_models(gvar_est, thin = 3)

}
