% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gvarmodel.R
\name{create_gvarmodel}
\alias{create_gvarmodel}
\title{Create a Global Vector Autoregressive Model}
\usage{
create_gvarmodel(submodel_data, global_data = NULL)
}
\arguments{
\item{submodel_data}{a named list of time-series objects of country-specific data.}

\item{global_data}{a named time-series object of global data.}
}
\value{
A list of class 'gvarmodel'.
}
\description{
Initialises an object of class 'gvarmodel', which contains all elements necessary
to set up, estimate and evaluate a Bayesian GVAR model.
}
\examples{

# Load data
data("gvar2023")
submodel_data <- gvar2023[["submodel_data"]] # Country series including weights
global_data <- gvar2023[["global_data"]] # Global commodities data

# Create 'gvarmodel' object
object <- create_gvarmodel(submodel_data = submodel_data,
                           global_data = global_data)




}
