% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gvecmodel.R
\name{create_gvecmodel}
\alias{create_gvecmodel}
\title{Create a Global Vector Error Correction Model}
\usage{
create_gvecmodel(submodel_data, global_data = NULL)
}
\arguments{
\item{submodel_data}{an object of class 'submodeldata', which is a named list
of time-series objects of with endogenous data for sub-models.}

\item{global_data}{a named time-series object of global data.}
}
\value{
A list of class 'gvecmodel', which consists of the following elements:
\describe{
  \item{\strong{global}}{A named list containing the data objects of the global model.}
  \item{\strong{weights}}{The list entry, where the sub-model weight matrices will go.
  This entry is empty after the execution of \code{create_gvecmodel}, but will be
  updated with \code{\link{add_weight_matrices}}.} 
  \item{\strong{submodels}}{The list entry, where the estimated sub-models will go.
  This entry is empty after the execution of \code{create_gvecmodel}, but will be
  updated with \code{\link{add_submodels}}.}
}
}
\description{
Initialises an object of class 'gvecmodel', which contains all elements necessary
to set up, estimate and evaluate a Bayesian GVEC model.
}
\details{
The function creates a list of class 'gvecmodel', which provides the basic
structure for the creation of sub-models, prior specification, initial value
generation, posterior simulation, model evaluation and structural analysis.
}
\examples{

# Load data
data("dees2007")
submodel_data <- dees2007[["submodel_data"]] # Country series including weights
global_data <- dees2007[["global_data"]] # Global commodities data

# Create 'gvarmodel' object
object <- create_gvecmodel(submodel_data = submodel_data,
                           global_data = global_data)
                           
}
