% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_models.R
\name{create_models}
\alias{create_models}
\title{Create Country Models}
\usage{
create_models(
  country_data,
  gvar_weights,
  global_data = NULL,
  model_specs = NULL
)
}
\arguments{
\item{country_data}{a named list of time-series objects of country-specific data.}

\item{gvar_weights}{a matrix or an array of weights, usually, the output of a call to \code{\link{create_weights}}.}

\item{global_data}{a named time-series object of global data.}

\item{model_specs}{a list of model specifications as produced by \code{\link{create_specifications}}.}
}
\value{
A list of country model input.
}
\description{
Produces a list of country models for each model specification that should be estimated.
}
\examples{
data("gvar2016")

country_data <- gvar2016$country_data
global_data <- gvar2016$global_data
region_weights <- gvar2016$region_weights
weight_data <- gvar2016$weight_data

# Take first difference of all variables y and Dp
country_data <- diff_variables(country_data, variables = c("y", "Dp"))

# Generate weight matrices as 2 year, rolling window averages
gvar_weights <- create_weights(weight_data = weight_data, period = 2,
                               country_data = country_data)

# Create an object with country model specifications
model_specs <- create_specifications(country_data = country_data,
                                     global_data = global_data,
                                     variables = c("y", "Dp", "poil"),
                                     p_domestic = 2,
                                     p_foreign = 2,
                                     type = "VAR")

# Create estimable objects
object <- create_models(country_data = country_data,
                        gvar_weights = gvar_weights,
                        global_data = global_data,
                        model_specs = model_specs)

}
