% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_regions.R
\name{create_regions}
\alias{create_regions}
\title{Create Regional Series}
\usage{
create_regions(submodel_data, region_weights, regions, period)
}
\arguments{
\item{submodel_data}{a named list of class 'submodeldata'.}

\item{region_weights}{a multivariate time-series object containing data used
to weight the observations in \code{submodel_data}.}

\item{regions}{a named list of character vectors containing specifications for countries
in \code{submodel_data}, which should be combined to a region. The name of the
respective list element will become the name of the region.}

\item{period}{either a single integer or a numeric vector specifiying the periods in
\code{region_weights}, which should be used to construct weights.}
}
\value{
An object of class 'submodeldata'.
}
\description{
Combines multiple country-specific time series to regional series.
}
\details{
If a numeric vector is provided for argument \code{period}, the function weights
submodel-specific observations based on the sums over the specified periods.
If an integer is proved, the submodel-specific observations are weighted
according to rolling sums over the last \code{period} periods. If submodel
data starts earlier than the series in \code{region_weights}, the sums over
the first \code{period} observations of \code{region_weights} are used
until the periods match.
}
\examples{
# Load data
data("gvar2019")

new_regions <- list("EA" = c("AT", "BE", "DE", "ES", "FI", "FR", "IT", "NL"))

# Create regions
submodel_data <- create_regions(submodel_data = gvar2019[["submodel_data"]],
                                region_weights = gvar2019[["region_weights"]],
                                regions = new_regions,
                                period = 3)

}
