% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_specifications.R
\name{create_specifications}
\alias{create_specifications}
\title{Create Country Model Specifications}
\usage{
create_specifications(
  country_data,
  global_data = NULL,
  countries = NULL,
  domestic = NULL,
  foreign = NULL,
  global = NULL,
  deterministic = NULL,
  type = "VAR",
  r = NULL,
  structural = NULL,
  tvp = FALSE,
  sv = FALSE,
  iterations = 50000,
  burnin = 5000
)
}
\arguments{
\item{country_data}{a named list of time-series objects of country-specific data.}

\item{global_data}{a named time-series object of global data.}

\item{countries}{a character vector of country names, which should enter the GVAR model.
The names must correspond to the names of the elements in \code{country_data}. If
\code{NULL} (default), all countries are included.}

\item{domestic}{a named list of specificatios for domestic variables. See also 'Details'.}

\item{foreign}{a named list of specificatios for foreign variables. See also 'Details'.}

\item{global}{a named list of specificatios for global variables. See also 'Details'.}

\item{deterministic}{a named list of specifications for deterministic terms. See alos 'Details'.}

\item{type}{a character specifying if the country models are estimated as vector
autoregressive "VAR" (default) or vector error correction "VEC" models.}

\item{r}{an integer or vector for the cointegration rank of VEC models. See also 'Details'.}

\item{structural}{a character vector of country names, for which structural models
should be estimated. If \code{NULL} (default), no country-model will be structural.}

\item{tvp}{logical indicating whether the coefficients of the model are time varying (default is \code{FALSE}).}

\item{sv}{logical indicating whether the error variances are time varying and should be estimated using a stochastic volatility algorithm (default is \code{FALSE}).}

\item{iterations}{an integer of MCMC draws excluding burn-in draws (defaults
to 50000).}

\item{burnin}{an integer of MCMC draws used to initialize the sampler
(defaults to 5000). These draws do not enter the computation of posterior
moments, forecasts etc.}
}
\value{
A list of model specifications for each country.
}
\description{
Produces a list of model specifications for each country in a GVAR model.
}
\details{
If arguments \code{domestic}, \code{foreign} or \code{global} are \code{NULL} (default), all available variables
are used and the lag order is set to 1 for each group of variables. If a list is provided, the used variables can be
changed by adding an element \code{"variables"} with a character vector, i.e. \code{variables = c("var1", "var2", ...)},
where \code{"var1"} is the name of a column in object \code{"country_data"}. Similarly, the lag order of each
variable group can be changed by adding an element \code{"lags"} with an integer or integer vector, i.e. \code{lags = 1}.
If a vector is provided, \code{\link{create_models}} will produce a distinct model for all
possible specifications.

If a list is provided as argument \code{deterministic}, it can contain the following elements:
\describe{
 \item{\code{const}}{either logical for VAR models or a character specifying whether a constant term
 enters the error correction term (\code{"restricted"}) or the non-cointegration term (\code{"unrestricted"})
 of a VEC model. If `NULL` (default) no constant term will be added.}
 \item{\code{trend}}{either logical for VAR models or a character specifying whether a trend term
 enters the error correction term (\code{"restricted"}) or the non-cointegration term (\code{"unrestricted"})
 of a VEC model. If `NULL` (default) no trend term will be added.}
 \item{\code{seasonal}}{either logical for VAR models or a character specifying whether seasonal dummies
 enter the error correction term (\code{"restricted"}) or the non-cointegreation term (\code{"unrestricted"}).
 If `NULL` (default) no seasonal terms will be added. The amount of dummy variables depends
 on the frequency of the time series data in each list element.}
}

The argument \code{rank} is only used when \code{type = "VEC"}. Otherwise, the rank is set to zero.
If a vector is provided, \code{\link{create_models}} will produce a distinct model for all possible specifications.
}
\examples{
# Load data
data("gvar2019")

# Create regions
temp <- create_regions(country_data = gvar2019$country_data,
             weight_data = gvar2019$weight_data,
             region_weights = gvar2019$region_weights,
             regions = list(EA =  c("AT", "BE", "DE", "ES", "FI", "FR", "IT", "NL")),
             period = 3)

country_data <- temp$country_data
weight_data <- temp$weight_data
global_data = gvar2019$global_data

# Difference series to make them stationary
country_data <- diff_variables(country_data, variables = c("y", "Dp", "r"), multi = 100)
global_data <- diff_variables(global_data, multi = 100)

# Create time varying weights
weight_data <- create_weights(weight_data, period = 3, country_data = country_data)

# Generate specifications
model_specs <- create_specifications(
                 country_data = country_data,
                 global_data = global_data,
                 countries = c("US", "JP", "CA", "NO", "GB", "EA"), 
                 domestic = list(variables = c("y", "Dp", "r"), lags = 1),
                 foreign = list(variables = c("y", "Dp", "r"), lags = 1),
                 global = list(variables = c("poil"), lags = 1),
                 deterministic = list(const = TRUE, trend = FALSE, seasonal = FALSE),
                 iterations = 10,
                 burnin = 10)
# Note that the number of iterations and burnin draws should be much higher!

}
