% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_varxsubmodel.R
\name{create_varxsubmodel}
\alias{create_varxsubmodel}
\title{Create Sub-Models}
\usage{
create_varxsubmodel(
  object,
  submodel,
  endogen = NULL,
  p_endogen = 1,
  exogen = NULL,
  p_exogen = 1,
  global = NULL,
  s = NULL,
  deterministic = "const",
  seasonal = FALSE,
  structural = FALSE,
  tvp = FALSE,
  error = "wishart",
  varsel = "none",
  iterations = 10000,
  burnin = 2000
)
}
\arguments{
\item{object}{an object of class 'gvarmodel' or 'gvecmodel'.}

\item{submodel}{name of the sub-model, for which input data should be generated
based on the input in argument \code{object}.}

\item{endogen}{character vector of variables that should enter each sub-model
as endogenous variables, if they are available.}

\item{p_endogen}{an integer vector of the lag order (default is \code{p_endogen = 1})
of a sub-model's endogenous variables.}

\item{exogen}{character vector of variables that should enter each sub-model
as weakly exogenous variables.}

\item{p_exogen}{an integer vector of the lag order (default is \code{p_exogen = 1})
of a sub-model's weakly exogenous variables.}

\item{global}{character vector of variables that should enter each sub-model
as global variables.}

\item{s}{an integer vector of the lag order of a sub-model's global variables.
If \code{NULL} (default), models do not include global variables.}

\item{deterministic}{a character specifying which deterministic terms should
be included. Available values are \code{"none"}, \code{"const"} (default) for an intercept,
\code{"trend"} for a linear trend, and \code{"both"} for an intercept with a linear trend.}

\item{seasonal}{logical. If \code{TRUE}, seasonal dummy variables are
generated as additional deterministic terms. The amount of dummies depends on the frequency of the
time-series object provided in \code{object}. Defaults to \code{FALSE}.}

\item{structural}{logical indicating whether data should be prepared for the estimation of a
structural VAR model. Defaults to \code{FALSE}.}

\item{tvp}{logical indicating whether the model parameters are time varying.
Defaults to \code{FALSE}.}

\item{error}{character specifying the model that should be used for the estimation
of the covariance matrix of the error term. Default is \code{"wishart"}. See 'Details'.}

\item{varsel}{character specifying the type of variable selection algorithm
that should be employed. Default is \code{"none"}. See 'Details'.}

\item{iterations}{an integer of MCMC draws excluding burn-in draws (defaults
to 10000).}

\item{burnin}{an integer of MCMC draws used to initialize the sampler
(defaults to 2000). These draws do not enter the computation of posterior
moments, forecasts etc.}
}
\value{
An object of class 'modellist', which contains at least one element
of class 'varxsubmodel'.
}
\description{
Produces a list of VARX models for each sub-model in a GVAR model.
}
\details{
If an integer vector is provided as argument \code{p_endogen}, \code{p_exogen}
or \code{s}, the function will produce a distinct model for all
possible combinations of those specifications.

Argument \code{error} specifies the structure of the covariance matrix of
the error term and how it is estimated. Possible specifications are:
\itemize{
 \item{\code{"wishart"}: The covariance is estimated using a Wishart prior.}
 \item{\code{"gamma"}: Only the diagonal elements of the covariance matrix
 are estimated using a gamma prior.
 Off-diagonal elements are not estimated and set to zero.}
 \item{\code{"gamma+covar"}: The diagonal elements of the covariance matrix
 are estimated using a gamma prior.
Covariances are estimated based on a triangular decomposition.}
 \item{\code{"sv"}: Only the diagonal elements of the covariance matrix are
 estimated using a stochastic volatility
algorithm. Off-diagonal elements are not estimated and set to zero.}
 \item{\code{"sv+covar"}: Only the diagonal elements of the covariance matrix
 are estimated using a stochastic volatility
algorithm. Covariances are estimated based on a triangular decomposition.}
}

Available specifications for argument \code{varsel} are:
\itemize{
 \item{\code{"none"}: No variable selection algorithm is used.}
 \item{\code{"bvs"}: Bayesian variable selection as proposed in Korobilis (2013).}
 \item{\code{"ssvs"}: Stochastic search variable selection as proposed in George et al. (2008).}
}
}
\examples{

# Load data
data("gvar2019")
global_data <- gvar2019[["global_data"]]
submodel_data <- gvar2019[["submodel_data"]]

# Limit number of sub-models
submodel_data <- select_list_elements(submodel_data, c("AT", "DE", "US"))

# Create global model
object <- create_gvarmodel(submodel_data = submodel_data,
                           global_data = global_data)

# Generate and add weight matrices
object <- add_weight_matrices(object = object,
                              submodel_data = submodel_data,
                              period = 2013:2016)

# Create sub-model
object <- create_varxsubmodel(object,
                              submodel = "AT",
                              endogen = c("y","Dp", "r"), p_endogen = 1,
                              exogen = c("y", "Dp"), p_exogen = 1,
                              global = "poil", s = 0,
                              deterministic = "const", seasonal = FALSE,
                              structural = FALSE, tvp = FALSE,
                              error = "wishart", varsel = "none",
                              iterations = 10, burnin = 10)
# Number of iterations and burn-in should be much higher.


}
