% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_vecxsubmodel.R
\name{create_vecxsubmodel}
\alias{create_vecxsubmodel}
\title{Create Sub-Models}
\usage{
create_vecxsubmodel(
  object,
  submodel,
  endogen = NULL,
  p_endogen = 1,
  exogen = NULL,
  p_exogen = 1,
  global = NULL,
  s = NULL,
  r = NULL,
  const = NULL,
  trend = NULL,
  seasonal = NULL,
  structural = FALSE,
  tvp = FALSE,
  error = "wishart",
  varsel = "none",
  algorithm = NULL,
  iterations = 10000,
  burnin = 2000
)
}
\arguments{
\item{object}{an object of class 'gvarmodel' or 'gvecmodel'.}

\item{submodel}{name of the sub-model, for which input data should be generated
based on the input in argument \code{object}.}

\item{endogen}{character vector of variables that should enter each sub-model
as endogenous variables, if they are available.}

\item{p_endogen}{an integer vector of the lag order (default is \code{p_endogen = 1})
of a sub-model's endogenous variables.}

\item{exogen}{character vector of variables that should enter each sub-model
as weakly exogenous variables.}

\item{p_exogen}{an integer vector of the lag order (default is \code{p_exogen = 1})
of a sub-model's weakly exogenous variables.}

\item{global}{character vector of variables that should enter each sub-model
as global variables.}

\item{s}{an integer vector of the lag order of a sub-model's global variables.
If \code{NULL} (default), models do not include global variables.}

\item{r}{an integer vector of the cointegration rank. See 'Details'.}

\item{const}{a character specifying whether a constant term enters the error correction
term (\code{"restricted"}) or the non-cointegration term as an \code{"unrestricted"} variable.
If \code{NULL} (default) no constant term will be added.}

\item{trend}{a character specifying whether a trend term enters the error correction
term (\code{"restricted"}) or the non-cointegration term as an \code{"unrestricted"} variable.
If \code{NULL} (default) no constant term will be added.}

\item{seasonal}{a character specifying whether seasonal dummies should be included in the error
correction term (\code{"restricted"}) or in the non-cointegreation term as \code{"unrestricted"}
variables. If \code{NULL} (default) no seasonal terms will be added. The amount of dummy variables
will be automatically detected and depends on the frequency of the time-series object provided
in \code{data}.}

\item{structural}{logical indicating whether data should be prepared for the estimation of a
structural VEC model. Defaults to \code{FALSE}.}

\item{tvp}{logical indicating whether the model parameters are time varying.
Defaults to \code{FALSE}.}

\item{error}{character specifying the model that should be used for the estimation
of the covariance matrix of the error term. Default is \code{"wishart"}. See 'Details'.}

\item{varsel}{character specifying the type of variable selection algorithm
that should be employed. Default is \code{"none"}. See 'Details'.}

\item{algorithm}{algorithm that should be used for posterior simulation. If \code{NULL}
(default), standard algorithms will be used. See 'Details' for available
non-standard options.}

\item{iterations}{an integer of MCMC draws excluding burn-in draws (defaults
to 10000).}

\item{burnin}{an integer of MCMC draws used to initialize the sampler
(defaults to 2000). These draws do not enter the computation of posterior
moments, forecasts etc.}
}
\value{
An object of class 'modellist', which contains at least one element
of class 'vecxsubmodel'.
}
\description{
Produces a list of VECX models for each sub-model in a GVEC model.
}
\details{
If an integer vector is provided as argument \code{p_endogen}, \code{p_exogen},
\code{s} or \code{r}, the function will produce a distinct model for all
possible combinations of those specifications.

Argument \code{error} specifies the structure of the covariance matrix of
the error term and how it is estimated. Possible specifications are:
\itemize{
 \item{\code{"wishart"}: The covariance is estimated using a Wishart prior.}
 \item{\code{"gamma"}: Only the diagonal elements of the covariance matrix
 are estimated using a gamma prior.
 Off-diagonal elements are not estimated and set to zero.}
 \item{\code{"gamma+covar"}: The diagonal elements of the covariance matrix
 are estimated using a gamma prior.
Covariances are estimated based on a triangular decomposition.}
 \item{\code{"sv"}: Only the diagonal elements of the covariance matrix are
 estimated using a stochastic volatility
algorithm. Off-diagonal elements are not estimated and set to zero.}
 \item{\code{"sv+covar"}: Only the diagonal elements of the covariance matrix
 are estimated using a stochastic volatility
algorithm. Covariances are estimated based on a triangular decomposition.}
}

Available specifications for argument \code{varsel} are:
\itemize{
 \item{\code{"none"}: No variable selection algorithm is used.}
 \item{\code{"bvs"}: Bayesian variable selection as proposed in Korobilis (2013).}
 \item{\code{"ssvs"}: Stochastic search variable selection as proposed in George et al. (2008).}
}

Available specifications for argument \code{algorithm} are:
\itemize{
   \item{\code{"KLGS2010"}: Algorithm proposed in Koop, León-González & Strachan (2010).}
}
}
\examples{

# Load data
data("dees2007")
submodel_data <- dees2007[["submodel_data"]]
global_data <- dees2007[["global_data"]]

# Limit number of sub-models
submodel_data <- select_list_elements(submodel_data, c("EA", "CA", "US"))

# Create empty model
object <- create_gvecmodel(submodel_data = submodel_data,
                           global_data = global_data)

# Add weight matrices
object <- add_weight_matrices(object = object,
                              submodel_data = submodel_data,
                              period = 1999:2001)

# Create sub-models
model <- create_vecxsubmodel(object,
                             submodel = "EA",
                             r = 1,
                             global = "poil",
                             s = 1,
                             iterations = 10,
                             burnin = 10)
# Number of iterations and burn-in should be much higher.


}
