% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dees2007}
\alias{dees2007}
\title{Dees, di Mauro, Pesaran \& Smith data set}
\format{
A named list containing the following elements:
\describe{
  \item{\strong{country_data}}{A named list of 26 time-series objects with 99 quarterly
  observations from 1979Q2 to 2003Q4 of the following variables:
  
  \itemize{
  \item \strong{y:} log real GDP
  \item \strong{p:} rate of inflation
  \item \strong{rs:} short-term interest rate
  \item \strong{rl:} long-term interest rate
  \item \strong{q:} log real equity prices
  \item \strong{ep:} log real exchange rate
  }
  }
  
  \item{\strong{global_data}}{A time-series object containing 99 quarterly observations
  from 1979Q2 to 2003Q4 of log of oil prices.} 
  \item{\strong{weight_data}}{A named \eqn{26 \times 26} weight matrix.}
  \item{\strong{dees_specs}}{A named list of country model specifications as used in
  Dees et al. (2007).}
}

See the supplementary material to Dees et al. (2007) for further details on the
sources and compilation of the data (\url{http://qed.econ.queensu.ca/jae/datasets/dees001/}).
}
\usage{
data("dees2007")
}
\description{
The data set of Dees et al. (2007) contains economic time series for
26 economies and oil prices from 1979Q2 to 2003Q4.
}
\references{
Dees, S., di Mauro, F., Pesaran, M. H., & Smith, V. (2007). Exploring the
international linkages of the euro area: A global VAR analysis.
\emph{Journal of Applied Econometrics, 22}(1), 1--38.
\url{https://doi.org/10.1002/jae.932}
}
\keyword{datasets}
