% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_variables.R
\name{diff_variables}
\alias{diff_variables}
\title{Differences of Variables}
\usage{
diff_variables(data, variables = NULL, multi = NULL)
}
\arguments{
\item{data}{a named time-series object or a list of named time-series objects.}

\item{variables}{a character vector of variables that should be differenced, if
they appear in a time-series object. If \code{NULL} (default), all variables are differenced.}

\item{multi}{optional. Numeric by which the differenced series should be multiplicated.}
}
\value{
A differenced time-series object or a list of differenced time-series objects.
}
\description{
Produces the first difference of variables in a time-series object or in a
list of named time-series objects.
}
\examples{
data("gvar2016")

country_data <- gvar2016$country_data

# Take first difference of the variables "y" and "Dp" across all
# elements of obect "country_data"
country_data <- diff_variables(country_data, variables = c("y", "Dp"))

}
