% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_submodel_specifications.R
\name{get_submodel_specifications}
\alias{get_submodel_specifications}
\title{Specifications of the Submodels of a GVAR Model}
\usage{
get_submodel_specifications(object)
}
\arguments{
\item{object}{a list of the estimated models of one sub-model, usually one
element of \code{object$submodels} after a call to
\code{\link[bvartools]{add_posterior_coefficients}}.}
}
\value{
A data frame.
}
\description{
Obtains the model specification of the submodels of a GVAR model.
}
