% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girf.R
\name{girf}
\alias{girf}
\title{Gernalised Impulse Response Function}
\usage{
girf(
  object,
  impulse,
  response,
  n.ahead = 5,
  shock = "sd",
  ci = 0.95,
  cumulative = FALSE,
  mc.cores = NULL
)
}
\arguments{
\item{object}{an object of class \code{"bgvar"}, usually, a result of a call to \code{\link{solve_gvar}}.}

\item{impulse}{a character vector of the impulse country and variable, respectively.}

\item{response}{a character vector of the response country and variable, respectively.y.}

\item{n.ahead}{number of steps ahead.}

\item{shock}{size of the shock. Either \code{"sd"} (default) for a standard deviation of
the impulse variable, \code{"nsd"} for a negativ standard deviation, or a numeric.}

\item{ci}{a numeric between 0 and 1 specifying the probability mass covered by the
credible intervals. Defaults to 0.95.}

\item{cumulative}{logical specifying whether a cumulative IRF should be calculated.}

\item{mc.cores}{the number of cores to use, i.e. at most how many child
processes will be run simultaneously. The option is initialized from
environment variable MC_CORES if set. Must be at least one, and
parallelization requires at least two cores.}
}
\value{
A time-series object of class \code{"bgvarirf"}.
}
\description{
Computes the generalised impulse response coefficients of an object of class \code{"bgvar"} for
`n.ahead` steps.
}
\references{
Lütkepohl, H. (2007). \emph{New introduction to multiple time series analyis} (2nd ed.). Berlin: Springer.

Pesaran, H. H., & Shin, Y. (1998). Generalized impulse response analysis in linear multivariate models. \emph{Economics Letters, 58}, 17-29.
}
