% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/girf.R
\name{girf}
\alias{girf}
\title{Gernalised Impulse Response Function}
\usage{
girf(
  object,
  impulse,
  response,
  n.ahead = 5,
  shock = "sd",
  ci = 0.95,
  cumulative = FALSE,
  mc.cores = NULL
)
}
\arguments{
\item{object}{an object of class \code{"bgvar"}, usually, a result of a call to \code{\link{solve_gvar}}.}

\item{impulse}{a character vector of the impulse country and variable, respectively.}

\item{response}{a character vector of the response country and variable, respectively.y.}

\item{n.ahead}{number of steps ahead.}

\item{shock}{size of the shock. Either \code{"sd"} (default) for a standard deviation of
the impulse variable, \code{"nsd"} for a negativ standard deviation, or a numeric.}

\item{ci}{a numeric between 0 and 1 specifying the probability mass covered by the
credible intervals. Defaults to 0.95.}

\item{cumulative}{logical specifying whether a cumulative IRF should be calculated.}

\item{mc.cores}{the number of cores to use, i.e. at most how many child
processes will be run simultaneously. The option is initialized from
environment variable MC_CORES if set. Must be at least one, and
parallelization requires at least two cores.}
}
\value{
A time-series object of class \code{"bgvarirf"}.
}
\description{
Computes the generalised impulse response coefficients of an object of class \code{"bgvar"} for
`n.ahead` steps.
}
\examples{
# Load data
data("gvar2016")

# Create regions
temp <- create_regions(country_data = gvar2016$country_data,
                       weight_data = gvar2016$weight_data,
                       region_weights = gvar2016$region_weights,
                       regions = list(EA =  c("AT", "BE", "DE", "ES", "FI", "FR", "IT", "NL")), period = 3)

country_data <- temp$country_data
weight_data <- temp$weight_data
global_data = gvar2016$global_data

# Difference series to make them stationary
country_data <- diff_variables(country_data, variables = c("y", "Dp", "r"), multi = 100)
global_data <- diff_variables(global_data, multi = 100)

# Create weights
weight_data <- create_weights(weight_data, period = 3, country_data = country_data)

# Create general model specifications
model_specs <- create_specifications(country_data = country_data,
                                     global_data = global_data,
                                     domestic = list(variables = c("y", "Dp", "r"), lags = 1:2),
                                     foreign = list(variables = c("y", "Dp", "r"), lags = 1:2),
                                     global = list(variables = "poil", lags = 1:2),
                                     deterministic = list("const" = TRUE),
                                     countries = c("EA", "US", "GB", "CA", "JP", "IN"),
                                     type = "VAR",
                                     tvp = FALSE, sv = FALSE,
                                     iterations = 10,
                                     burnin = 10)
# Note that the number of iterations and burnin should be much higher!
                                     
# Country-specific specifications
model_specs[["US"]][["domestic"]][["variables"]] <- c("y", "Dp", "r")
model_specs[["US"]][["foreign"]][["variables"]] <- c("y", "Dp")

# Create all country models
country_models <- create_models(country_data = country_data,
                                weight_data = weight_data,
                                global_data = global_data,
                                model_specs = model_specs)


# Add priors
temp_model <- add_priors(country_models,
                         coef = list(v_i = 1 / 9, v_i_det = 1 / 10),
                         sigma = list(df = 3, scale = 1))

# Obtain posterior draws
object <- draw_posterior(temp_model)

# Solve GVAR
gvar <- combine_submodels(object)

# Obtain GIRF
gvar_girf <- girf(gvar, impulse = c("EA", "r"), response = c("EA", "y"), n.ahead = 20)

# Plot GIRF
plot(gvar_girf)

}
\references{
Lütkepohl, H. (2007). \emph{New introduction to multiple time series analyis} (2nd ed.). Berlin: Springer.

Pesaran, H. H., & Shin, Y. (1998). Generalized impulse response analysis in linear multivariate models. \emph{Economics Letters, 58}, 17-29.
}
