% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inclusion_prior.varxsubmodel.R
\name{inclusion_prior.varxsubmodel}
\alias{inclusion_prior.varxsubmodel}
\title{Prior Inclusion Probabilities}
\usage{
\method{inclusion_prior}{varxsubmodel}(
  object,
  prob = 0.5,
  exclude_deterministics = TRUE,
  minnesota_like = FALSE,
  kappa1 = 0.8,
  kappa2 = 0.5,
  kappa3 = 0.5,
  kappa4 = 0.8,
  ...
)
}
\arguments{
\item{object}{an object of class \code{"varxsubmodel"}, usually, a result of a
call to \code{\link{create_varxsubmodel}}.}

\item{prob}{a numeric specifying the prior inclusion probability of all model parameters.}

\item{exclude_deterministics}{logical. If \code{TRUE} (default), the vector of the positions of
included variables does not include the positions of deterministic terms.}

\item{minnesota_like}{logical. If \code{TRUE}, the prior inclusion probabilities of the
parameters are calculated in a similar way as the Minnesota prior. See 'Details'.}

\item{kappa1}{a numeric specifying the prior inclusion probability of
coefficients that correspond to own lags of endogenous variables.
Only used if \code{minnesota_like = TRUE}. See 'Details'.}

\item{kappa2}{a numeric specifying the size of the prior inclusion probabilities
of endogenous variables, which do not correspond to own lags.
Only used if \code{minnesota_like = TRUE}. See 'Details'.}

\item{kappa3}{a numeric specifying the size of the prior inclusion probabilities
of non-deterministic exogenous variables. Default is \code{NULL}, which indicates that the formula
for the calculation of the prior inclusion probabilities of deterministic terms
is used for all exogenous variables.
Only used if \code{minnesota_like = TRUE}. See 'Details'.}

\item{kappa4}{a numeric specifying the size of the prior inclusion probabilities
of deterministic terms. Only used if \code{minnesota_like = TRUE}. See 'Details'.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list containing a matrix of prior inclusion probabilities and an integer vector
specifying the positions of variables, which should be included in the variable selection algorithm.
}
\description{
Prior inclusion probabilities as required for stochastic search variable selection (SSVS) à la
George et al. (2008) and Bayesian variable selection (BVS) à la Korobilis (2013).
}
\details{
If \code{minnesota_like = TRUE}, prior inclusion probabilities \eqn{\underline{\pi}_1}
are calculated as
\tabular{cl}{
\eqn{\frac{\kappa_1}{r}} \tab for own lags of endogenous variables, \cr
\eqn{\frac{\kappa_2}{r}} \tab for other endogenous variables, \cr
\eqn{\frac{\kappa_3}{1 + r}} \tab for unmodelled exogenous variables, \cr
\eqn{\kappa_{4}} \tab for deterministic variables.
}
}
\examples{

# Load data
data("gvar2019")
global_data <- gvar2019[["global_data"]]
submodel_data <- gvar2019[["submodel_data"]]

# Limit number of sub-models
submodel_data <- select_list_elements(submodel_data, c("AT", "DE", "US"))

# Create global model
object <- create_gvarmodel(submodel_data = submodel_data,
                           global_data = global_data)

# Generate and add weight matrices
object <- add_weight_matrices(object = object,
                              submodel_data = submodel_data,
                              period = 2013:2016)

# Create sub-model
object <- create_varxsubmodel(object,
                              submodel = "AT",
                              endogen = c("y","Dp", "r"), p_endogen = 1,
                              exogen = c("y", "Dp"), p_exogen = 2,
                              global = "poil", s = 1,
                              deterministic = "const", seasonal = FALSE,
                              structural = FALSE, tvp = FALSE,
                              error = "wishart", varsel = "none",
                              iterations = 10, burnin = 10)
# Number of iterations and burn-in should be much higher.

# The previous function returns a model list. Extract the first model to proceed.
object <- object[[1]]

prior <- inclusion_prior(object)

}
