% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.varxsubmodelest.R
\name{irf.varxsubmodelest}
\alias{irf.varxsubmodelest}
\title{Impulse Response Function for a GVAR Submodel}
\usage{
\method{irf}{varxsubmodelest}(
  x,
  impulse,
  response,
  n_ahead = 5,
  ci = 0.95,
  shock = 1,
  type = "feir",
  cumulative = FALSE,
  keep_draws = FALSE,
  period = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class 'varxsubmodelest', usually, a result of a call to
\code{\link{draw_posterior.gvarsubmodels}} or \code{\link[bvartools]{bvec_to_bvar}}.}

\item{impulse}{name of the impulse variable.}

\item{response}{name of the response variable.}

\item{n_ahead}{number of steps ahead.}

\item{ci}{a numeric between 0 and 1 specifying the probability mass covered by the
credible intervals. Defaults to 0.95.}

\item{shock}{size of the shock.}

\item{type}{type of the impulse response. Possible choices are forecast error \code{"feir"}
(default), orthogonalised \code{"oir"}, structural \code{"sir"}, generalised \code{"gir"},
and structural generalised \code{"sgir"} impulse responses.}

\item{cumulative}{logical specifying whether a cumulative IRF should be calculated.}

\item{keep_draws}{logical specifying whether the function should return all draws of
the posterior impulse response function. Defaults to \code{FALSE} so that
the median and the credible intervals of the posterior draws are returned.}

\item{period}{integer. Index of the period, for which the IR should be generated.
Only used for TVP or SV models. Default is \code{NULL}, so that the posterior draws of the last time period
are used.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A time-series object of class \code{"bvarirf"} and if \code{keep_draws = TRUE} a simple matrix.
}
\description{
Computes the impulse response coefficients of an object of class 'varxsubmodelest' for
\code{n_ahead} steps.
}
\references{
Lütkepohl, H. (2006). \emph{New introduction to multiple time series analysis} (2nd ed.). Berlin: Springer.

Pesaran, H. H., Shin, Y. (1998). Generalized impulse response analysis in linear multivariate models. \emph{Economics Letters, 58}, 17-29.
}
