% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bgvecest.R
\name{plot.bgvecest}
\alias{plot.bgvecest}
\title{Plotting Draws of a VECX Submodel of a GVAR Model}
\usage{
\method{plot}{bgvecest}(x, ci = 0.95, type = "hist", model = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{"bgvecest"}, usually, a result of a call to \code{\link{draw_posterior}}.}

\item{ci}{interval used to calculate credible bands for time-varying parameters.}

\item{type}{either \code{"hist"} (default) for histograms, \code{"trace"} for a trace plot,
or \code{"boxplot"} for a boxplot. Only used for parameter draws of constant coefficients.}

\item{model}{indicates, for which model plots should be produced. Can be a character
vector of country names or an integer vector of the positions of the elements in
argument \code{x}.}

\item{...}{further graphical parameters.}
}
\description{
A plot function for objects of class \code{"bgvecest"}.
}
