% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.submodelestlist.R
\name{plot.submodelestlist}
\alias{plot.submodelestlist}
\title{Plotting Draws of Submodels of a GVAR Model}
\usage{
\method{plot}{submodelestlist}(x, ci = 0.95, type = "hist", variables = "all", group = NULL, ...)
}
\arguments{
\item{x}{an object of class 'submodelestlist', usually, a result of a call to
\code{\link[bvartools]{draw_posterior}}.}

\item{ci}{interval used to calculate credible bands for time-varying parameters.}

\item{type}{either \code{"hist"} (default) for histograms, \code{"trace"} for a trace plot,
or \code{"boxplot"} for a boxplot. Only used for parameter draws of constant coefficients.}

\item{variables}{character vector of variables that should be plotted. Default is \code{"all"}.
Other options are \code{"domestic"}, \code{"foreign"}, \code{"global"}, \code{"deterministic"}
and \code{"sigma"}.}

\item{group}{character. Name of the element in argument \code{x}, for which posterior draws
should be plotted. If \code{NULL} (default), all submodels are used.}

\item{...}{further graphical parameters.}
}
\description{
A plot function for objects of class 'submodelestlist'.
}
