% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.submodelselcritlist.R
\name{plot.submodelselcritlist}
\alias{plot.submodelselcritlist}
\title{Plotting Selection Criteria}
\usage{
\method{plot}{submodelselcritlist}(x, criterion = "BIC", ...)
}
\arguments{
\item{x}{an object of class 'submodelselcritlist', usually, a result of a call
to \code{\link{selection_criteria}}.}

\item{criterion}{the selection criterion that should be plotted. Available choices
are \code{"LL"}, \code{"AIC"}, \code{"BIC"} (default), \code{"HQ"}.}

\item{...}{further graphical parameters passed on to \link[graphics]{boxplot}.}
}
\description{
A plot function for objects of class 'submodelselcritlist'.
}
