% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.vecxsubmodelest.R
\name{plot.vecxsubmodelest}
\alias{plot.vecxsubmodelest}
\title{Plotting Draws of a VECX Submodel of a GVAR Model}
\usage{
\method{plot}{vecxsubmodelest}(x, ci = 0.95, type = "hist", variables = "all", ...)
}
\arguments{
\item{x}{an object of class \code{"vecxsubmodelest"}, usually, a result of a call to \code{\link[bvartools]{add_posterior_coefficients}}.}

\item{ci}{interval used to calculate credible bands for time-varying parameters.}

\item{type}{either \code{"hist"} (default) for histograms, \code{"trace"} for a trace plot,
or \code{"boxplot"} for a boxplot. Only used for parameter draws of constant coefficients.}

\item{variables}{character vector of variables that should be plotted. Default is \code{"all"}.
Other options are \code{"domestic"}, \code{"foreign"}, \code{"global"}, \code{"deterministic"}
and \code{"sigma"}.}

\item{...}{further graphical parameters.}
}
\description{
A plot function for objects of class \code{"vecxsubmodelest"} for visual inspection
of posterior draws.
}
