% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.bgvar.R
\name{predict.bgvar}
\alias{predict.bgvar}
\title{Predict Method for Objects of Class bgvar}
\usage{
\method{predict}{bgvar}(object, ..., n.ahead = 10, new_x = NULL, new_d = NULL, ci = 0.95)
}
\arguments{
\item{object}{an object of class "bgvar", usually, a result of a call to \code{\link{combine_models}}.}

\item{...}{additional arguments.}

\item{n.ahead}{number of steps ahead at which to predict.}

\item{new_x}{a matrix of new non-deterministic, exogenous variables. Must have \code{n.ahead} rows.}

\item{new_d}{a matrix of new deterministic variables. Must have \code{n.ahead} rows.}

\item{ci}{a numeric between 0 and 1 specifying the probability mass covered by the
credible intervals. Defaults to 0.95.}
}
\value{
A time-series object of class "bgvarprd".
}
\description{
Forecasting a Bayesian Global VAR object of class "bgvar" with credible bands.
}
\details{
The function produces \code{n.ahead} forecasts for the GVAR model
\deqn{y_t = \sum_{l = 1}^{p} G_{l} y_{t-i} + \sum_{l = 0}^{s} H_{l} x_{t-i} + D d_t + G^{-1}_{0} u_t,}
with \eqn{u_t \sim N(0, \Sigma)}.
}
\references{
Lütkepohl, H. (2007). \emph{New introduction to multiple time series analyis}. Berlin: Springer.
}
