% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.bgvar.R
\name{predict.bgvar}
\alias{predict.bgvar}
\title{Predict Method for Objects of Class bgvar}
\usage{
\method{predict}{bgvar}(object, ..., n.ahead = 10, new_x = NULL, new_D = NULL, ci = 0.95)
}
\arguments{
\item{object}{an object of class "bgvar", usually, a result of a call to \code{\link{solve_gvar}}.}

\item{...}{additional arguments.}

\item{n.ahead}{number of steps ahead at which to predict.}

\item{new_x}{a matrix of new non-deterministic, exogenous variables. Must have \code{n.ahead} rows.}

\item{new_D}{a matrix of new deterministic variables. Must have \code{n.ahead} rows.}

\item{ci}{a numeric between 0 and 1 specifying the probability mass covered by the
credible intervals. Defaults to 0.95.}
}
\value{
A time-series object of class "bgvarprd".
}
\description{
Forecasting a Bayesian Global VAR object of class "bgvar" with credible bands.
}
\details{
The function produces \code{n.ahead} forecasts for the GVAR model
\deqn{y_t = \sum_{l = 1}^{p} G_{l} y_{t-i} + \sum_{l = 0}^{s} H_{l} x_{t-i} + D d_t + G^{-1}_{0} u_t,}
with \eqn{u_t \sim N(0, \Sigma)}.
}
\examples{
data("gvar2016")

country_data <- gvar2016$country_data
global_data <- gvar2016$global_data
region_weights <- gvar2016$region_weights
weight_data <- gvar2016$weight_data

# Take first difference of all variables y and Dp
country_data <- diff_variables(country_data, variables = c("y", "Dp", "r"))

# Generate EA area region with 2 year rolling window weights
ea <- c("AT", "BE", "DE", "ES", "FI", "FR", "IT", "NL")
temp <- create_regions(country_data = country_data,
                       regions = list("EA" = ea),
                       period = 2,
                       region_weights = region_weights,
                       weight_data = weight_data)
country_data <- temp$country_data
weight_data <- temp$weight_data

# Generate weight matrices as 2 year, rolling window averages
gvar_weights <- create_weights(weight_data = weight_data, period = 2,
                               country_data = country_data)

# Create an object with country model specifications
model_specs <- create_specifications(country_data = country_data,
                                     global_data = global_data,
                                     variables = c("y", "Dp", "r"),
                                     countries = c("EA", "US", "JP", "CA", "MX", "GB"),
                                     p_domestic = 1,
                                     p_foreign = 1,
                                     type = "VAR")

# Create estimable objects
object <- create_models(country_data = country_data,
                        gvar_weights = gvar_weights,
                        model_specs = model_specs)

# Add priors
object <- add_priors(object)

# Estimate GVAR model
gvar_est <- estimate_gvar(object, iterations = 100, burnin = 10, thin = 2)
# Note that the number of iterations and burnin should be much higher.

# Solve GVAR
gvar_solved <- solve_gvar(gvar_est)

# Forecasts
gvar_predict <- predict(gvar_solved)

# Plot Forecasts
plot(gvar_predict, variable = c("US", "Dp"))

}
\references{
Lütkepohl, H. (2007). \emph{New introduction to multiple time series analyis}. Berlin: Springer.
}
