% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_submodel_from_folder.R
\name{read_submodel_from_folder}
\alias{read_submodel_from_folder}
\title{Import Sub-Models from a Folder of HDF5 Files}
\usage{
read_submodel_from_folder(folder, submodel)
}
\arguments{
\item{folder}{Path to the root folder of a global model. Usually, the same
path used as argument \code{"folder"} during a call to
\code{\link{write_to_hdf5.gvarmodel}}.}

\item{submodel}{character of the sub-model that should be imported.}
}
\value{
A list of class 'modellist'.
}
\description{
Imports every model stored below a sub-model's folder.
}
\details{
The function allows to import the data from a single sub-model of a global
model. It's main purpose is to avoid reading all models of a global model at
once in order to save memory.
}
\examples{

# Load data
data("gvar2023")
submodel_data <- gvar2023[["submodel_data"]]
global_data <- gvar2023[["global_data"]]

# Limit number of sub-models
submodel_data <- select_list_elements(submodel_data, c("AT", "DE", "US"))

# Set up a model
object <- create_gvarmodel(submodel_data = submodel_data,
                           global_data = global_data)
                           
# Generate and add weight matrices
object <- add_weight_matrices(object = object,
                              submodel_data = submodel_data,
                              period = 2014:2016)
                              
# Create sub-models
object <- add_submodels(object,
                        endogen = c("y", "Dp", "r"),
                        p_endogen = 1,
                        exogen = c("y", "Dp", "r"),
                        p_exogen = 1,
                        global = "poil",
                        s = 1,
                        r = 1,
                        error = "wishart",
                        iterations = 10,
                        burnin = 10)
# Number of iterations and burn-in should be much higher.

# Add priors
object <- add_priors(object,
                     coef = list(v_i = 0),
                     coint = list(v_i = 0, p_tau_i = 1),
                     sigma = list(df = 3, scale = 0.0001))
                     
# Add initial values
object <- add_initial_values(object)

# Export models
folder <- file.path(tempdir(), "gvar")
dir.create(folder)
write_to_hdf5(object, folder = folder)

# Import
submodel <- read_submodel_from_folder(folder = folder, submodel = "AT")


}
