% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_submodels.R
\name{select_submodels}
\alias{select_submodels}
\title{Susbmodel Selection}
\usage{
select_submodels(object, ic = "BIC", select = "order", teststats = NULL)
}
\arguments{
\item{object}{a list containing the results of the country estimates, usually, the
result of a call to \code{\link{draw_posterior}}.}

\item{ic}{a character specifying the information criterion used for model selection.
Available options are "AIC", "BIC" (default) and "HQ".}

\item{select}{a character specifying how the best country model is selected.
If \code{"order"} (default), the country model with the overall minimum value of the
specified information criterion per country is selected. If \code{"rank"}, the
function selects the model, after which the selected information criterion increases
for the first time.}

\item{teststats}{optional. A named list of test statistics for each model in `object`. Usually,
the result of a call to \code{\link{submodel_test_statistics}}.}
}
\description{
Selects the best country-specific submodel of a GVAR model.
}
\examples{
# Load data
data("gvar2019")

# Create regions
temp <- create_regions(country_data = gvar2019$country_data,
             weight_data = gvar2019$weight_data,
             region_weights = gvar2019$region_weights,
             regions = list(EA =  c("AT", "BE", "DE", "ES", "FI", "FR", "IT", "NL")),
             period = 3)

country_data <- temp$country_data
weight_data <- temp$weight_data
global_data = gvar2019$global_data

# Difference series to make them stationary
country_data <- diff_variables(country_data, variables = c("y", "Dp", "r"), multi = 100)
global_data <- diff_variables(global_data, multi = 100)

# Create time varying weights
weight_data <- create_weights(weight_data, period = 3, country_data = country_data)

# Generate specifications
model_specs <- create_specifications(
                 country_data = country_data,
                 global_data = global_data,
                 countries = c("US", "JP", "CA", "NO", "GB", "EA"), 
                 domestic = list(variables = c("y", "Dp", "r"), lags = 1:2),
                 foreign = list(variables = c("y", "Dp", "r"), lags = 1:2),
                 global = list(variables = c("poil"), lags = 1:2),
                 deterministic = list(const = TRUE, trend = FALSE, seasonal = FALSE),
                 iterations = 10,
                 burnin = 10)
# Note that the number of iterations and burnin draws should be much higher!
                                     
# Overwrite country-specific specifications
model_specs[["US"]][["domestic"]][["variables"]] <- c("y", "Dp", "r")
model_specs[["US"]][["foreign"]][["variables"]] <- c("y", "Dp")

# Create estimation objects
country_models <- create_models(country_data = country_data,
                                weight_data = weight_data,
                                global_data = global_data,
                                model_specs = model_specs)

# Add priors
models_with_priors <- add_priors(country_models,
                                 coef = list(v_i = 1 / 9, v_i_det = 1 / 10),
                                 sigma = list(df = 3, scale = .0001))

# Obtain posterior draws
object <- draw_posterior(models_with_priors)

# Select final submodels
object <- select_submodels(object, ic = "BIC", select = "order")


}
