% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.gvarmodel.R
\name{window.gvarmodel}
\alias{window.gvarmodel}
\title{Time Series Windows}
\usage{
\method{window}{gvarmodel}(x, start = NULL, end = NULL, ...)
}
\arguments{
\item{x}{an object of class 'gvarmodel'.}

\item{start}{the start time of the period of interest.}

\item{end}{the end time of the period of interest.}
}
\value{
An object of class 'gvarmodel'.
}
\description{
Forwards model input to the same function for individual models.
}
