% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_methods.R
\name{window.gvecmodel}
\alias{window.gvecmodel}
\title{Time Series Windows}
\usage{
\method{window}{gvecmodel}(x, start = NULL, end = NULL, ...)
}
\arguments{
\item{x}{an object of class 'gvecmodel'.}

\item{start}{the start time of the period of interest.}

\item{end}{the end time of the period of interest.}
}
\value{
An object of class 'gvecmodel'.
}
\description{
Forwards model input to the same function for individual models.
}
